Control.Monad.Freer.Church

Description

The church-encoded Freer Monad. Basically provides the free monad in a way that is compatible with HFunctor and Interpret, and :.: or Compose in a way that is compatible with HBifunctor and Tensor and Monoidal.

Documentation

newtype Free f a

A Free f is f enhanced with "sequential binding" capabilities. It allows you to sequence multiple fs one after the other, and also to determine "what f to sequence" based on the result of the computation so far.

Essentially, you can think of this as "giving f a Monad instance", with all that that entails (return, >>=, etc.).

Lift f into it with inject :: f a -> Free f a. When you finally want to "use" it, you can interpret it into any monadic context:

interpret
    :: Monad g
    => (forall x. f x -> g x)
    -> Free f a
    -> g a

Under the hood, this is the Church-encoded Freer monad. It's Free, or F, but in a way that is compatible with HFunctor and Interpret.

Constructors

Free

Fields

runFree :: forall r. (a -> r) -> (forall s. f s -> (s -> r) -> r) -> r

reFree :: (MonadFree f m, Functor f) => Free f a -> m a

Convert a Free f into any instance of MonadFree f.

liftFree :: f ~> Free f

interpretFree :: Monad g => (f ~> g) -> Free f ~> g

retractFree :: Monad f => Free f ~> f

hoistFree :: (f ~> g) -> Free f ~> Free g

newtype Free1 f a

Constructors

Free1

Fields

runFree1 :: forall r. (forall s. f s -> (s -> a) -> r) -> (forall s. f s -> (s -> r) -> r) -> r

toFree :: Free1 f ~> Free f

liftFree1 :: f ~> Free1 f

free1Comp :: Free1 f ~> Comp f (Free f)

matchFree1 :: Functor f => Free1 f ~> (f :+: Comp f (Free1 f))

interpretFree1 :: Bind g => (f ~> g) -> Free1 f ~> g

retractFree1 :: Bind f => Free1 f ~> f

hoistFree1 :: (f ~> g) -> Free1 f ~> Free1 g

data Comp f g a

Functor composition. Comp f g a is equivalent to f (g a), and the Comp pattern synonym is a way of getting the f (g a) in a Comp f g a.

For example, Maybe (IO Bool) is Comp Maybe IO Bool.

This is mostly useful for its typeclass instances: in particular, Functor, Applicative, HBifunctor, and Monoidal.

This is essentially a version of :.: and Compose that allows for an HBifunctor instance.

It is slightly less performant. Using comp . unComp every once in a while will concretize a Comp value (if you have Functor f) and remove some indirection if you have a lot of chained operations.

The "free monoid" over Comp is Free, and the "free semigroup" over Comp is Free1.

Constructors

(f x) :>>= (x -> g a)

Bundled Patterns

pattern Comp :: Functor f => f (g a) -> Comp f g a
Pattern match on and construct a Comp f g a as if it were f (g a).

comp :: f (g a) -> Comp f g a

"Smart constructor" for Comp that doesn't require Functor f.