Data.Functor.Apply.Free

Documentation

data Ap1 :: (Type -> Type) -> Type -> Type where

The free Apply. Basically a "non-empty" Ap.

The construction here is based on Ap, similar to now NonEmpty is built on list.

Constructors

Ap1 :: f a -> Ap f (a -> b) -> Ap1 f b

toAp :: Ap1 f ~> Ap f

An Ap1 is a "non-empty" Ap; this function "forgets" the non-empty property and turns it back into a normal Ap.

liftAp1 :: f ~> Ap1 f

Embed an f into Ap1.

retractAp1 :: Apply f => Ap1 f ~> f

Extract the f out of the Ap1.

retractAp1 . liftAp1 == id

runAp1 :: Apply g => (f ~> g) -> Ap1 f ~> g

Interpret an Ap f into some Apply context g.