Data.Functor.Associative

Associative

class HBifunctor t => Associative t where

An HBifunctor where it doesn't matter which binds first is Associative. Knowing this gives us a lot of power to rearrange the internals of our HFunctor at will.

For example, for the functor product:

data (f :*: g) a = f a :*: g a

We know that f :*: (g :*: h) is the same as (f :*: g) :*: h.

Methods

associating :: (Functor f, Functor g, Functor h) => t f (t g h) <~> t (t f g) h

The isomorphism between t f (t g h) a and t (t f g) h a. To use this isomorphism, see assoc and disassoc.

assoc :: (Associative t, Functor f, Functor g, Functor h) => t f (t g h) ~> t (t f g) h

Reassociate an application of t.

disassoc :: (Associative t, Functor f, Functor g, Functor h) => t (t f g) h ~> t f (t g h)

Reassociate an application of t.

Semigroupoidal

class (Associative t, Interpret (SF t)) => Semigroupoidal t where

Minimal complete definition

appendSF, matchSF

Associated Types

type SF t :: (Type -> Type) -> Type -> Type

Methods

appendSF :: t (SF t f) (SF t f) ~> SF t f

If a SF t f represents multiple applications of t f to itself, then we can also "append" two SF t fs applied to themselves into one giant SF t f containing all of the t fs.

matchSF :: Functor f => SF t f ~> (f :+: t f (SF t f))

consSF :: t f (SF t f) ~> SF t f

Prepend an application of t f to the front of a SF t f.

toSF :: t f f ~> SF t f

Embed a direct application of f to itself into a SF t f.

biretract :: C (SF t) f => t f f ~> f

A version of retract that works for a Tensor. It retracts both fs into a single f.

binterpret :: C (SF t) h => (f ~> h) -> (g ~> h) -> t f g ~> h

A version of interpret that works for a Tensor. It takes two interpreting functions, and interprets both joined functors one after the other into h.

matchingSF :: (Semigroupoidal t, Functor f) => SF t f <~> (f :+: t f (SF t f))

An SF t f represents the successive application of t to f, over and over again. So, that means that an SF t f must either be a single f, or an t f (SF t f).

matchingSF states that these two are isomorphic. Use matchSF and inject !*! consSF to convert between one and the other.

Utility

extractT :: (Semigroupoidal t, C (SF t) f, Copointed f) => t f f a -> a

Useful wrapper over retractT to allow you to directly extract an a from a t f f a, if f is a valid retraction from t, and f is an instance of Copointed.

Useful fs include Identity or related newtype wrappers from base:

extractT
    :: (Monoidal t, C (MF t) Identity)
    => t Identity Identity a
    -> a

getT :: (Semigroupoidal t, C (SF t) (Const b)) => (forall x. f x -> b) -> (forall x. g x -> b) -> t f g a -> b

Useful wrapper over interpret to allow you to directly extract a value b out of the t f a, if you can convert f x into b.

Note that depending on the constraints on the interpretation of t, you may have extra constraints on b.

If C (MF t) is Unconstrained, there are no constraints on b

If C (MF t) is Apply, b needs to be an instance of Semigroup

If C (MF t) is Applicative, b needs to be an instance of Monoid

For some constraints (like Monad), this will not be usable.

-- Return the length of either the list, or the Map, depending on which
--   one s in the +
length !*! length
    :: ([] :+: Map Int) Char
    -> Int

-- Return the length of both the list and the map, added together
(Sum . length) !*! (Sum . length)
    :: Day [] (Map Int) Char
    -> Sum Int

collectT :: (Semigroupoidal t, C (SF t) (Const [b])) => (forall x. f x -> b) -> (forall x. g x -> b) -> t f g a -> [b]

Useful wrapper over getT to allow you to collect a b from all instances of f and g inside a t f g a.

This will work if C t is Unconstrained, Apply, or Applicative.

(!*!) :: (Semigroupoidal t, C (SF t) h) => (f ~> h) -> (g ~> h) -> t f g ~> h infixr 5

Infix alias for binterpret

(!$!) :: (Semigroupoidal t, C (SF t) (Const b)) => (forall x. f x -> b) -> (forall x. g x -> b) -> t f g a -> b infixr 5

Infix alias for getT

Chain1

data Chain1 t f a

Constructors

Done1 (f a)

More1 (t f (Chain1 t f) a)

foldChain1 :: forall t f g. HBifunctor t => (f ~> g) -> (t f g ~> g) -> Chain1 t f ~> g

unfoldChain1 :: forall t f g. HBifunctor t => (g ~> (f :+: t f g)) -> g ~> Chain1 t f

unrollingSF :: forall t f. (Semigroupoidal t, Functor f) => SF t f <~> Chain1 t f

A type SF t is supposed to represent the successive application of ts to itself. The type Chain1 t f is an actual concrete ADT that contains successive applications of t to itself, and you can pattern match on each layer.

unrollingSF states that the two types are isormorphic. Use runollSF and rerollSF to convert between the two.

unrollSF :: forall t f. (Semigroupoidal t, Functor f) => SF t f ~> Chain1 t f

A type SF t is supposed to represent the successive application of ts to itself. unrollSF makes that successive application explicit, buy converting it to a literal Chain1 of applications of t to itself.

rerollSF :: Semigroupoidal t => Chain1 t f ~> SF t f

A type SF t is supposed to represent the successive application of ts to itself. rerollSF takes an explicit Chain1 of applications of t to itself and rolls it back up into an SF t.