Data.Functor.Combinator

Description

Functor combinators and tools (typeclasses and utiility functions) to manipulate them.

Classes include:

HFunctor and HBifunctor, used to swap out the functors that the combinators modify

Interpret, Tensor, Monoidal, used to inject and interpret functor values with respect to their combinators.

We have some helpful utility functions, as well, built on top of these typeclasses.

The second half of this module exports the various useful functor combinators that can modify functors to add extra functionality, or join two functors together and mix them in different ways. Use them to build your final structure by combining simpler ones in composable ways!

See README for a tutorial and a rundown on each different functor combinator.

Classes

A lot of type signatures are stated in terms of ~>. ~> represents a "natural transformation" between two functors: a value of type f ~> g is a value of type 'f a -> g a that works for any a@.

type (~>) (f :: k -> Type) (g :: k -> Type) = forall (x :: k). f x -> g x infixr 0

A natural transformation from f to g.

Single Functors

Classes that deal with single-functor combinators, that enhance a single functor.

class HFunctor t where

An HFunctor can be thought of a unary "functor transformer" --- a basic functor combinator. It takes a functor as input and returns a functor as output.

It "enhances" a functor with extra structure (sort of like how a monad transformer enhances a Monad with extra structure).

As a uniform inteface, we can "swap the underlying functor" (also sometimes called "hoisting"). This is what hmap does: it lets us swap out the f in a t f for a t g.

For example, the free monad Free takes a Functor and returns a new Functor. In the process, it provides a monadic structure over f. hmap lets us turn a Free f into a Free g: a monad built over f can be turned into a monad built over g.

For the ability to move in and out of the enhanced functor, see Interpret.

Methods

hmap :: (f ~> g) -> t f ~> t g

If we can turn an f into a g, then we can turn a t f into a t g.

It must be the case that

hmap id == id

Essentially, t f adds some "extra structure" to f. hmap must swap out the functor, without affecting the added structure.

For example, ListF f a is essentially a list of f as. If we hmap to swap out the f as for g as, then we must ensure that the "added structure" (here, the number of items in the list) remains the same. So, hmap must preserve the number of items in the list.

The law hmap id == id is a way of formalizing this informal property.

class Inject t => Interpret t where

An Interpret lets us move in and out of the "enhanced" Functor.

For example, Free f is f enhanced with monadic structure. We get:

inject    :: f a -> Free f a
interpret :: Monad m => (forall x. f x -> m x) -> Free f a -> m a

inject will let us use our f inside the enhanced Free f. interpret will let us "extract" the f from a Free f if we can give an interpreting function that interprets f into some target Monad.

The type family C tells us the typeclass constraint of the "target" functor. For Free, it is Monad, but for other Interpret instances, we might have other constraints.

We enforce that:

interpret id . inject == id

That is, if we lift a value into our structure, then immediately interpret it out as itself, it should lave the value unchanged.

Minimal complete definition

retract | interpret

Associated Types

type C t :: (Type -> Type) -> Constraint

The constraint on the target context of interpret.

Methods

retract :: C t f => t f ~> f

Remove the f out of the enhanced t f structure, provided that f satisfies the necessary constraints. If it doesn't, it needs to be properly interpreted out.

interpret :: C t g => (f ~> g) -> t f ~> g

Given an "interpeting function" from f to g, interpret the f out of the t f into a final context g.

interpretFor :: (Interpret t, C t g) => t f a -> (f ~> g) -> g a

A convenient flipped version of interpret.

extractI :: (Interpret t, C t f, Copointed f) => t f a -> a

Useful wrapper over retract to allow you to directly extract an a from a t f a, if f is a valid retraction from t, and f is an instance of Copointed.

Useful fs include Identity or related newtype wrappers from base:

extractI
    :: (Interpret t, C t Identity)
    => t Identity a
    -> a

getI :: (Interpret t, C t (Const b)) => (forall x. f x -> b) -> t f a -> b

Useful wrapper over interpret to allow you to directly extract a value b out of the t f a, if you can convert f x into b.

Note that depending on the constraints on the interpretation of t, you may have extra constraints on b.

If C t is Unconstrained, there are no constraints on b

If C t is Apply, b needs to be an instance of Semigroup

If C t is Applicative, b needs to be an instance of Monoid

For some constraints (like Monad), this will not be usable.

-- get the length of the Map String in the Step.
collectI length
     :: Step (Map String) Bool
     -> [Int]

collectI :: (Interpret t, C t (Const [b])) => (forall x. f x -> b) -> t f a -> [b]

Useful wrapper over getI to allow you to collect a b from all instances of f inside a t f a.

This will work if C t is Unconstrained, Apply, or Applicative.

-- get the lengths of all Map Strings in the Ap.
collectI length
     :: Ap (Map String) Bool
     -> [Int]

Multi-Functors

Classes that deal with two-functor combinators, that "mix" two functors together in some way.

class HBifunctor t where

A HBifunctor is like an HFunctor, but it enhances two different functors instead of just one.

Usually, it enhaces them "together" in some sort of combining way.

This typeclass provides a uniform instance for "swapping out" or "hoisting" the enhanced functors. We can hoist the first one with hleft, the second one with hright, or both at the same time with hbimap.

For example, the f :*: g type gives us "both f and g":

data (f :*: g) a = f a :*: g a

It combines both f and g into a unified structure --- here, it does it by providing both f and g.

The single law is:

hbimap id id == id

This ensures that hleft, hright, and hbimap do not affect the structure that t adds on top of the underlying functors.

Minimal complete definition

hleft, hright | hbimap

Methods

hleft :: (f ~> j) -> t f g ~> t j g

Swap out the first transformed functor.

hright :: (g ~> k) -> t f g ~> t f k

Swap out the second transformed functor.

hbimap :: (f ~> j) -> (g ~> k) -> t f g ~> t j k

Swap out both transformed functors at the same time.

class Associative t => Tensor t

A HBifunctor can be a Tensor if:

There is some identity i where t i f is equivalent to just f. That is, "enhancing" f with t i does nothing.

t is associative: f t (g t h) is equivalent to (f t g) t h).

The methods in this class provide us useful ways of navigating a Tensor t with respect to this property.

Realistically, there won't be any Tensor instances that are not also Monoidal instances. The two classes are separated only to help organize functionality into cleaner sub-divisions.

Minimal complete definition

intro1, intro2, elim1, elim2

Associated Types

type I t :: Type -> Type

The identity of Tensor t. If you "combine" f with the identity, it leaves f unchanged.

For example, the identity of :*: is Proxy. This is because

(Proxy :*: f) a

is equivalent to just

f a

:*:-ing f with Proxy gives you no additional structure.

inL :: forall t f g a. (Monoidal t, C (MF t) g) => f a -> t f g a

Convenient wrapper over intro1 that lets us introduce an arbitrary functor g to the right of an f.

inR :: forall t f g a. (Monoidal t, C (MF t) f) => g a -> t f g a

Convenient wrapper over intro2 that lets us introduce an arbitrary functor f to the right of a g.

(!$!) :: (Semigroupoidal t, C (SF t) (Const b)) => (forall x. f x -> b) -> (forall x. g x -> b) -> t f g a -> b infixr 5

Infix alias for getT

extractT :: (Semigroupoidal t, C (SF t) f, Copointed f) => t f f a -> a

Useful wrapper over retractT to allow you to directly extract an a from a t f f a, if f is a valid retraction from t, and f is an instance of Copointed.

Useful fs include Identity or related newtype wrappers from base:

extractT
    :: (Monoidal t, C (MF t) Identity)
    => t Identity Identity a
    -> a

getT :: (Semigroupoidal t, C (SF t) (Const b)) => (forall x. f x -> b) -> (forall x. g x -> b) -> t f g a -> b

Useful wrapper over interpret to allow you to directly extract a value b out of the t f a, if you can convert f x into b.

Note that depending on the constraints on the interpretation of t, you may have extra constraints on b.

If C (MF t) is Unconstrained, there are no constraints on b

If C (MF t) is Apply, b needs to be an instance of Semigroup

If C (MF t) is Applicative, b needs to be an instance of Monoid

For some constraints (like Monad), this will not be usable.

-- Return the length of either the list, or the Map, depending on which
--   one s in the +
length !*! length
    :: ([] :+: Map Int) Char
    -> Int

-- Return the length of both the list and the map, added together
(Sum . length) !*! (Sum . length)
    :: Day [] (Map Int) Char
    -> Sum Int

(!*!) :: (Semigroupoidal t, C (SF t) h) => (f ~> h) -> (g ~> h) -> t f g ~> h infixr 5

Infix alias for binterpret

collectT :: (Semigroupoidal t, C (SF t) (Const [b])) => (forall x. f x -> b) -> (forall x. g x -> b) -> t f g a -> [b]

Useful wrapper over getT to allow you to collect a b from all instances of f and g inside a t f g a.

This will work if C t is Unconstrained, Apply, or Applicative.

Combinators

Functor combinators ** Single

data Coyoneda (f :: Type -> Type) a where

A covariant Functor suitable for Yoneda reduction

Constructors

Coyoneda :: forall (f :: Type -> Type) a b. (b -> a) -> f b -> Coyoneda f a

newtype ListF f a

A list of f as. Can be used to describe a product of many different values of type f a.

This is the Free Plus.

Constructors

ListF

Fields

runListF :: [f a]

newtype NonEmptyF f a

A non-empty list of f as. Can be used to describe a product between many different possible values of type f a.

Essentially:

NonEmptyF f
    ~ f                          -- one f
  :+: (f :*: f)              -- two f's
  :+: (f :*: f :*: f)            -- three f's
  :+: (f :*: f :*: f :*: f)      -- four f's
  :+: ...                        -- etc.

This is the Free Plus.

Constructors

NonEmptyF

Fields

runNonEmptyF :: NonEmpty (f a)

Bundled Patterns

pattern ProdNonEmpty :: (f :*: ListF f) a -> NonEmptyF f a
Treat a NonEmptyF f as a product between an f and a ListF f.

nonEmptyProd is the record accessor.

newtype MaybeF f a

A maybe f a.

Can be useful for describing a "an f a that may or may not be there".

This is the free structure for a "fail"-like typeclass that only has zero :: f a.

Constructors

MaybeF

Fields

runMaybeF :: Maybe (f a)

data Ap (f :: Type -> Type) a

The free Applicative for a Functor f.

data Ap1 :: (Type -> Type) -> Type -> Type

One or more fs convolved with itself.

Essentially:

Ap1 f
    ~ f                            -- one f
  :+: (f `'Day'\\` f)          -- two f's
  :+: (f `Day` f `Day` f)           -- three f's
  :+: (f `Day` f `Day` f `Day` f)  -- four f's
  :+: ...                          -- etc.

Useful if you want to promote an f to a situation with "at least one f sequenced with itself".

Mostly useful for its HFunctor and Interpret instance, along with its relationship with Ap and Day.

This is the free Apply --- Basically a "non-empty" Ap.

The construction here is based on Ap, similar to now NonEmpty is built on list.

data Alt (f :: Type -> Type) a

data Free f a

A Free f is f enhanced with "sequential binding" capabilities. It allows you to sequence multiple fs one after the other, and also to determine "what f to sequence" based on the result of the computation so far.

Essentially, you can think of this as "giving f a Monad instance", with all that that entails (return, >>=, etc.).

Lift f into it with inject :: f a -> Free f a. When you finally want to "use" it, you can interpret it into any monadic context:

interpret
    :: Monad g
    => (forall x. f x -> g x)
    -> Free f a
    -> g a

Under the hood, this is the Church-encoded Freer monad. It's Free, or F, but in a way that is compatible with HFunctor and Interpret.

data Step f a

An f a, along with a Natural index.

Step f a ~ (Natural, f a)
Step f   ~ ((,) Natural) :.: f       -- functor composition

It is the fixed point of applications of :+: (functor sums).

Intuitively, in an infinite f :+: f :+: f :+: f ..., you have exactly one f somewhere. A Step f a has that f, with a Natural giving you "where" the f is in the long chain.

Can be useful for using with the Monoidal instance of :+:.

interpreting it requires no constraint on the target context.

Constructors

Step

Fields

stepPos :: Natural

stepVal :: f a

newtype Steps f a

A non-empty map of Natural to f a. Basically, contains multiple f as, each at a given Natural index.

Steps f a ~ Map Natural (f a)
Steps f   ~ Map Natural :.: f       -- functor composition

It is the fixed point of applications of TheseT.

Intuitively, in an infinite f `TheseT` f `TheseT` f `TheseT` f ..., each of those infinite positions may have an f in them. However, because of the at-least-one nature of TheseT, we know we have at least one f at one position somewhere.

A Steps f a has potentially many fs, each stored at a different Natural position, with the guaruntee that at least one f exists.

Can be useful for using with the Monoidal instance of TheseT.

interpreting it requires at least an Alt instance in the target context, since we have to handle potentially more than one f.

Constructors

Steps

Fields

getSteps :: NEMap Natural (f a)

data ProxyF f a

The functor combinator that forgets all structure in the input. Ignores the input structure and stores no information.

Acts like the "zero" with respect to functor combinator composition.

ComposeT ProxyF f      ~ ProxyF
ComposeT f      ProxyF ~ ProxyF

It can be injected into (losing all information), but it is impossible to ever retract or interpret it.

Constructors

ProxyF

data Void2 a b

Void2 a b is uninhabited for all a and b.

Multi

data Day (f :: Type -> Type) (g :: Type -> Type) a where

The Day convolution of two covariant functors.

Constructors

Day :: forall (f :: Type -> Type) (g :: Type -> Type) a b c. f b -> g c -> (b -> c -> a) -> Day f g a

data ((f :: k -> Type) :*: (g :: k -> Type)) (p :: k) :: forall k. (k -> Type) -> (k -> Type) -> k -> Type infixr 6

Products: encode multiple arguments to constructors

Constructors

(f p) :*: (g p) infixr 6

data ((f :: k -> Type) :+: (g :: k -> Type)) (p :: k) :: forall k. (k -> Type) -> (k -> Type) -> k -> Type infixr 5

Sums: encode choice between constructors

Constructors

L1 (f p)

R1 (g p)

data Void1 a

Void1 is a functor that is uninhabited for all inputs. That is, Void1 a is uninhabited for all a.

The identity functor of :+: (and also TheseT)

data These1 f g a

A These f g a has either an f a, a g a, or both.

This is re-defined here from the these package, to avoid the high dependency footprint.

Constructors

This1 (f a)

That1 (g a)

These1 (f a) (g a)

data Comp f g a where

Functor composition. Comp f g a is equivalent to f (g a), and the Comp pattern synonym is a way of getting the f (g a) in a Comp f g a.

For example, Maybe (IO Bool) is Comp Maybe IO Bool.

This is mostly useful for its typeclass instances: in particular, Functor, Applicative, HBifunctor, and Monoidal.

This is essentially a version of :.: and Compose that allows for an HBifunctor instance.

It is slightly less performant. Using comp . unComp every once in a while will concretize a Comp value (if you have Functor f) and remove some indirection if you have a lot of chained operations.

The "free monoid" over Comp is Free, and the "free semigroup" over Comp is Free1.

Bundled Patterns

pattern Comp :: Functor f => f (g a) -> Comp f g a
Pattern match on and construct a Comp f g a as if it were f (g a).

newtype Final c f a

A simple way to inject/reject into any eventual typeclass.

In a way, this is the "ultimate" multi-purpose Interpret instance. You can use this to inject an f into a free structure of any typeclass. If you want f to have a Monad instance, for example, just use

inject :: f a -> Final Monad f a

When you want to eventually interpret out the data, use:

interpret :: (f ~> g) -> Final c f a -> g a

Essentially, Final c is the "free c". Final Monad is the free Monad, etc.

Final can theoretically replace Ap, Ap1, ListF, NonEmptyF, MaybeF, Free, Identity, Coyoneda, and other instances of FreeOf, if you don't care about being able to pattern match on explicit structure.

However, it cannot replace Interpret instances that are not free structures, like Step, Steps, Backwards, etc.

Note that this doesn't have instances for all the typeclasses you could lift things into; you probably have to define your own if you want to use Final c as an instance of c (using liftFinal0, liftFinal1, liftFinal2 for help).

Constructors

Final

Fields

runFinal :: forall g. c g => (forall x. f x -> g x) -> g a

class Interpret t => FreeOf c t | t -> c where

A typeclass associating a free structure with the typeclass it is free on.

This essentially lists instances of Interpret where a "trip" through Final will leave it unchanged.

fromFree . FreeOf == id
FreeOf . fromFree == id

This can be useful because Final doesn't have a concrete structure that you can pattern match on and inspect, but t might. This lets you work on a concrete structure if you desire.

Minimal complete definition

Nothing

Methods

fromFree :: t f ~> Final c f

toFree :: Functor f => Final c f ~> t f

fromFree :: C t (Final c f) => t f ~> Final c f

toFree :: c (t f) => Final c f ~> t f