Data.Functor.Combinator.Cons

Documentation

data Cons :: ((Type -> Type) -> Type -> Type -> Type) -> (Type -> Type) -> (Type -> Type) -> Type -> Type where

A lot of tensors can be factored out into a Cons on some parameterized profunctor. For example, we have:

:.: = Cons Star
Day = Cons Static
:*: = Cons ProdK

In the case of :.:, this actually allows us to have HFunctor and Interpret instances for functor composition, because the representation of :.: normally does not allow this. This gives us a proper tensor to form Free, which makes the free monad work nicely within our system.

In the normal case, you can just use Day and :*: instead of DayCons = Cons Static and ProdCons = Cons ProdK. The only one that gives us extra power is Comp = Cons Star.

You can use DayCons and ProdCons as Day and :*: using the provided pattern synonyms.

Bundled Patterns

pattern Comp :: Functor f => f (g a) -> Comp f g a

pattern DayCons :: Functor g => Day f g a -> DayCons f g a

pattern ProdCons :: Functor f => (f :*: g) a -> ProdCons f g a

class Monoidal (Cons p) => Uncons p t | t -> p where

Associates each Tensor t with its Cons equivalent, Cons p.

t need not always be a legal Tensor.

Methods

uncons :: Functor g => t f g ~> Cons p f g

cons :: Functor f => Cons p f g ~> t f g

Instances

Monadic

type Comp = Cons Star

A version of :.: that permits HBifunctor, Tensor, and Monoidal instances.

unComp :: Comp f g a -> Functor f -> f (g a)

comp :: f (g a) -> Comp f g a

Applicative

type DayCons = Cons Static

unDayCons :: DayCons b c ~> Day b c

Product

data ProdK g a b

Constructors

ProdK (a -> b) (g b)

type ProdCons = Cons ProdK

unProdCons :: ProdCons f g a -> Functor f -> (f :*: g) a

prodCons :: (f :*: g) ~> ProdCons f g