Data.Functor.Contravariant.Divise

Description

The contravariant counterpart of Apply: like Divisible, but without conquer. This is only a part of this library currently for compatibility, until it is (hopefully) merged into semigroupoids.

Since: 0.3.0.0

Documentation

class Contravariant f => Divise f where

The contravariant analogue of Apply; it is Divisible without conquer.

If one thinks of f a as a consumer of as, then divise allows one to handle the consumption of a value by splitting it between two consumers that consume separate parts of a.

divise takes the "splitting" method and the two sub-consumers, and returns the wrapped/combined consumer.

All instances of Divisible should be instances of Divise with divise = divide.

The guarantee that a function polymorphic over of Divise f provides that Divisible f doesn't that any input consumed will be passed to at least one sub-consumer; it won't potentially disappear into the void, as is possible if conquer is available.

Mathematically, a functor being an instance of Divise means that it is "semgroupoidal" with respect to the contravariant (tupling) Day convolution. That is, it is possible to define a function (f Day f) a -> f a in a way that is associative.

Methods

divise :: (a -> (b, c)) -> f b -> f c -> f a

Takes a "splitting" method and the two sub-consumers, and returns the wrapped/combined consumer.

divised :: Divise f => f a -> f b -> f (a, b)

Combine a consumer of a with a consumer of b to get a consumer of (a, b).

divised = divise id

(<:>) :: Divise f => f a -> f a -> f a

The Contravariant version of <|>: split the same input over two different consumers.

dsum1 :: (Foldable1 t, Divise f) => t (f a) -> f a

Convenient helper function to build up a Divise by splitting input across many different f as. Most useful when used alongside contramap:

dsum1 $ contramap get1 x
   :| [ contramap get2 y
      , contramap get3 z
      ]

Since: 0.3.3.0

newtype WrappedDivisible f a

Wrap a Divisible to be used as a member of Divise

Constructors

WrapDivisible

Fields

unwrapDivisible :: f a