Data.Functor.Contravariant.Divisible.Free

Description

Provides free structures for the various typeclasses of the Divisible hierarchy.

Since: 0.3.0.0

Documentation

newtype Div f a

The free Divisible. Used to sequence multiple contravariant consumers, splitting out the input across all consumers.

This type is essentially ListF; the only reason why it has to exist separately outside of ListF is because the current typeclass hierarchy isn't compatible with both the covariant Interpret instance (requiring Plus) and the contravariant Interpret instance (requiring Divisible).

The wrapping in Coyoneda is also to provide a usable Associative instance for the contravariant Day.

Constructors

Div

Fields

unDiv :: [Coyoneda f a]

Bundled Patterns

pattern Conquer :: Div f a
Pattern matching on an empty Div.

Before v0.3.3.0, this used to be the concrete constructor of Div. After, it is now an abstract pattern.

pattern Divide :: (a -> (b, c)) -> f b -> Div f c -> Div f a
Pattern matching on a non-empty Div, exposing the raw f instead of having it wrapped in a Coyoneda. This is the analogue of Pure and essentially treats the "cons" of the Div as a contravariant day convolution.

Before v0.3.3.0, this used to be the concrete constructor of Div. After, it is now an abstract pattern.

hoistDiv :: forall f g. (f ~> g) -> Div f ~> Div g

Map over the undering context in a Div.

liftDiv :: f ~> Div f

Inject a single action in f into a Div f.

runDiv :: forall f g. Divisible g => (f ~> g) -> Div f ~> g

Interpret a Div into a context g, provided g is Divisible.

divListF :: forall f. Contravariant f => Div f ~> ListF f

Div is isomorphic to ListF for contravariant f. This witnesses one way of that isomorphism.

listFDiv :: ListF f ~> Div f

Div is isomorphic to ListF for contravariant f. This witnesses one way of that isomorphism.

newtype Div1 f a

The free Divise: a non-empty version of Div.

This type is essentially NonEmptyF; the only reason why it has to exist separately outside of NonEmptyF is because the current typeclass hierarchy isn't compatible with both the covariant Interpret instance (requiring Plus) and the contravariant Interpret instance (requiring Divisible).

The wrapping in Coyoneda is also to provide a usable Associative instance for the contravariant Day.

Constructors

Div1

Fields

unDiv1 :: NonEmpty (Coyoneda f a)

Bundled Patterns

pattern Div1_ :: (a -> (b, c)) -> f b -> Div f c -> Div1 f a
Pattern matching on a Div1, exposing the raw f instead of having it wrapped in a Coyoneda. This is the analogue of Ap1 and essentially treats the "cons" of the Div1 as a contravariant day convolution.

Before v0.3.3.0, this used to be the concrete constructor of Div1. After, it is now an abstract pattern.

Since: 0.3.3.0

hoistDiv1 :: (f ~> g) -> Div1 f ~> Div1 g

Map over the underlying context in a Div1.

liftDiv1 :: f ~> Div1 f

Inject a single action in f into a Div f.

toDiv :: Div1 f ~> Div f

A Div1 is a "non-empty" Div; this function "forgets" the non-empty property and turns it back into a normal Div.

runDiv1 :: Divise g => (f ~> g) -> Div1 f ~> g

Interpret a Div1 into a context g, provided g is Divise.

div1NonEmptyF :: Contravariant f => Div1 f ~> NonEmptyF f

Div1 is isomorphic to NonEmptyF for contravariant f. This witnesses one way of that isomorphism.

nonEmptyFDiv1 :: NonEmptyF f ~> Div1 f

Div1 is isomorphic to NonEmptyF for contravariant f. This witnesses one way of that isomorphism.

data Dec :: (Type -> Type) -> Type -> Type where

The free Decide. Used to aggregate multiple possible consumers, directing the input into an appropriate consumer.

Constructors

Lose :: (a -> Void) -> Dec f a

Choose :: (a -> Either b c) -> f b -> Dec f c -> Dec f a

hoistDec :: forall f g. (f ~> g) -> Dec f ~> Dec g

Map over the underlying context in a Dec.

liftDec :: f ~> Dec f

Inject a single action in f into a Dec f.

runDec :: forall f g. Conclude g => (f ~> g) -> Dec f ~> g

Interpret a Dec into a context g, provided g is Conclude.

data Dec1 :: (Type -> Type) -> Type -> Type where

The free Decide: a non-empty version of Dec.

Constructors

Dec1 :: (a -> Either b c) -> f b -> Dec f c -> Dec1 f a

hoistDec1 :: forall f g. (f ~> g) -> Dec1 f ~> Dec1 g

Map over the undering context in a Dec1.

liftDec1 :: f ~> Dec1 f

Inject a single action in f into a Dec1 f.

toDec :: Dec1 f a -> Dec f a

A Dec1 is a "non-empty" Dec; this function "forgets" the non-empty property and turns it back into a normal Dec.

runDec1 :: Decide g => (f ~> g) -> Dec1 f ~> g

Interpret a Dec1 into a context g, provided g is Decide.