Data.Functor.Contravariant.Night

Description

Provides Night, a form of the day convolution that is contravariant and splits on Either.

Since: 0.3.0.0

Documentation

data Night :: (Type -> Type) -> (Type -> Type) -> Type -> Type where

A pairing of contravariant functors to create a new contravariant functor that represents the "choice" between the two.

A Night f g a is a contravariant "consumer" of a, and it does this by either feeding the a to f, or feeding the a to g. Which one it gives it to happens at runtime depending what a is actually given.

For example, if we have x :: f a (a consumer of as) and y :: g b (a consumer of bs), then night x y :: Night f g (Either a b). This is a consumer of Either a bs, and it consumes Left branches by feeding it to x, and Right branches by feeding it to y.

Mathematically, this is a contravariant day convolution, except with a different choice of bifunctor (Either) than the typical one we talk about in Haskell (which uses (,)). Therefore, it is an alternative to the typical Day convolution --- hence, the name Night.

Constructors

Night :: f b -> g c -> (a -> Either b c) -> Night f g a

night :: f a -> g b -> Night f g (Either a b)

Inject into a Night.

night x y is a consumer of Either a b; Left will be passed to x, and Right will be passed to y.

runNight :: Decide h => (f ~> h) -> (g ~> h) -> Night f g ~> h

Interpret out of a Night into any instance of Decide by providing two interpreting functions.

necide :: Decide f => Night f f ~> f

Squash the two items in a Night using their natural Decide instances.

Since: 0.4.0.0

assoc :: Night f (Night g h) ~> Night (Night f g) h

Night is associative.

unassoc :: Night (Night f g) h ~> Night f (Night g h)

Night is associative.

swapped :: Night f g ~> Night g f

The two sides of a Night can be swapped.

trans1 :: (f ~> h) -> Night f g ~> Night h g

Hoist a function over the left side of a Night.

trans2 :: (g ~> h) -> Night f g ~> Night f h

Hoist a function over the right side of a Night.

intro1 :: g ~> Night Not g

The left identity of Night is Not; this is one side of that isomorphism.

intro2 :: f ~> Night f Not

The right identity of Night is Not; this is one side of that isomorphism.

elim1 :: Contravariant g => Night Not g ~> g

The left identity of Night is Not; this is one side of that isomorphism.

elim2 :: Contravariant f => Night f Not ~> f

The right identity of Night is Not; this is one side of that isomorphism.

newtype Not a

A value of type Not a is "proof" that a is uninhabited.

Constructors

Not

Fields

refute :: a -> Void

refuted :: Not Void

A useful shortcut for a common usage: Void is always not so.

Since: 0.3.1.0