Data.Functor.Tensor

Description

There are two ways to write a Monoidal instance:

Define nilTM, consTM, unconsTM, and appendTM. These allow you to manipulate TM t as if it were a "list" of fs appended together with t.

Define fromF, toF, and appendF. F is a special data type that literally represents a linked list of fs appended together with t. The default definitions of nilTM, consTM, etc. then work on this representation.

Additionally, this class contains retractT, interpretT, pureT, and toTM. These are useful functions of using t as an interpreter combinator. They can all be derived from other methods, but they are provided as a part of the typeclass to allow implementors to provide more efficient versions.

Documentation

class HBifunctor t where

A HBifunctor is like an HFunctor, but it enhances two different functors instead of just one.

Usually, it enhaces them "together" in some sort of combining way.

This typeclass provides a uniform instance for "swapping out" or "hoisting" the enhanced functors. We can hoist the first one with hleft, the second one with hright, or both at the same time with hbimap.

For example, the f :*: g type gives us "both f and g":

data (f :*: g) a = f a :*: g a

It combines both f and g into a unified structure --- here, it does it by providing both f and g.

The single law is:

hbimap id id == id

This ensures that hleft, hright, and hbimap do not affect the structure that t adds on top of the underlying functors.

Minimal complete definition

hleft, hright | hbimap

Methods

hleft :: (f ~> j) -> t f g ~> t j g

Swap out the first transformed functor.

hright :: (g ~> k) -> t f g ~> t f k

Swap out the second transformed functor.

hbimap :: (f ~> j) -> (g ~> k) -> t f g ~> t j k

Swap out both transformed functors at the same time.

class HBifunctor t => Tensor t where

A HBifunctor can be a Tensor if:

There is some identity i where t i f is equivalent to just f. That is, "enhancing" f with t i does nothing.

t is associative: f t (g t h) is equivalent to @(f t g) t h).

The methods in this class provide us useful ways of navigating a Tensor t with respect to this property.

Associated Types

type I t :: Type -> Type

The identity of Tensor t. If you "combine" f with the identity, it leaves f unchanged.

For example, the identity of :*: is Proxy. This is because

(Proxy :*: f) a

is equivalent to just

f a

:*:-ing f with Proxy gives you no additional structure.

Methods

intro1 :: f ~> t f (I t)

intro2 :: Functor g => g ~> t (I t) g

elim1 :: Functor f => t f (I t) ~> f

elim2 :: Functor g => t (I t) g ~> g

assoc :: (Functor f, Functor g, Functor h) => t f (t g h) ~> t (t f g) h

disassoc :: (Functor f, Functor g, Functor h) => t (t f g) h ~> t f (t g h)

reconsTM :: forall t f. Monoidal t => (I t :+: t f (TM t f)) ~> TM t f

The inverse of unconsTM. Calls nilTM on the left (nil) branch, and consTM on the right (cons) branch.

class (Tensor t, Interpret (TM t)) => Monoidal t where

For some tensors t, you can represt the act of repeatedly combining the same functor an arbitrary amount of times:

t f f                    -- 2 times
t f (t f f)              -- 3 times
t f (t f (t f f))        -- 4 times
t f (t f (t f (t f f)))  -- 5 times

Sometimes, we have a type that can describe this repeated combination. For example, ListF f is the type that contains f :*:d with itself many number of times, and Ap is the type that contains f Dayd with itself many number of times.

ListF [x, y]       == x :*: y
ListF [x, y, z]    == x :*: y :*: z
ListF [x, y, z, q] == x :*: y :*: z :*: q

This is convenient because it allows you to represent repeated applications of t as a single data type.

For example, Day f f can be interpreted as "two sequenced fs", allowing you to specify "I want exactly two sequenced fs". If you want to specify "I want 0, 1, or many fs sequenced after each other", then you can use Ap f.

And, f :*: f can be interpreted as "a free selection of two fs", allowing you to specify "I have to fs that I can use". If you want to specify "I want 0, 1, or many different fs that I can use", you can use ListF f.

The Monoidal class unifies different such patterns. The associated type TM is the "repeated aplications of t" type.

See documentation of Data.Functor.Tensor for information on how to define instances of this typeclass.

Minimal complete definition

nilTM, consTM, unconsTM, appendTM | fromF, toF, appendF

Associated Types

type TM t :: (Type -> Type) -> Type -> Type

Methods

nilTM :: I t ~> TM t f

If TM t f represents multiple applications of t f with itself, then nilTM gives us "zero applications of f".

Note that t cannot be inferred from the type of nilTM, so this function must always be called with -XTypeApplications:

nilTM @Day :: Identity ~> Ap f
nilTM @Comp :: Identity ~> Free f
nilTM @(:*:) :: Proxy ~> ListF f

Together with consTM, forms an inverse with unconsTM.

consTM :: t f (TM t f) ~> TM t f

Prepend an application of t f to the front of a TM t f.

Together with nilTM, forms an inverse with unconsTM.

unconsTM :: TM t f ~> (I t :+: t f (TM t f))

If a TM t f represents multiple applications of t f to itself, unconsTM lets us break it up into two possibilities:

The TM t f had no applications of f

The TM t f had at least one application of f; we return the "first" f applied to the rest of the fs.

Should form an inverse with reconsTM:

reconsTM . unconsTM == id
unconsTM . reconsTM == id

where reconsTM is nilTM on the left side of the :+:, and consTM on the right side of the :+:.

appendTM :: t (TM t f) (TM t f) ~> TM t f

If a TM t f represents multiple applications of t f to itself, then we can also "append" two TM t fs applied to themselves into one giant TM t f containing all of the t fs.

fromF :: F t (I t) f ~> TM t f

Convert a linked list of t fs applied to themselves (stored in the F type) into TM t f, the data type representing multiple applications of t f to itself.

F i (I t) can be thought of as a "universal" representation of multiple-applications-to-self, and TM t can be thought of as a tailor-made represenation for your specific Tensor t.

fromF . toF == id
toF . fromF == id

toF :: TM t f ~> F t (I t) f

The inverse of fromF: convert a TM t f into a linked list of t fs applied to themselves. See fromF for more information.

appendF :: t (F t (I t) f) (F t (I t) f) ~> F t (I t) f

Append two linked lists of t f applied to itself together.

retractT :: C (TM t) f => t f f ~> f

A version of retract that works for a Tensor. It retracts both fs into a single f.

interpretT :: C (TM t) h => (f ~> h) -> (g ~> h) -> t f g ~> h

A version of interpret that works for a Tensor. It takes two interpreting functions, and interprets both joined functors one after the other into h.

pureT :: C (TM t) f => I t ~> f

If we have an instance of t, we can generate an f based on how it interacts with t.

Specialized (and simplified), this type is:

pureT @Day   :: Applicative f => a -> f a  -- pure
pureT @Comp  :: Monad f => a -> f a    -- return
pureT @(:*:) :: Plus f => f a          -- zero

toTM :: t f f ~> TM t f

Embed a direct application of f to itself into a TM t f.

data F t i f a

A useful construction that works like a "linked list" of t f applied to itself multiple times. That is, it contains t f f, t f (t f f), t f (t f (t f f)), etc.

If t is Monoidal, then it means we can "collapse" this linked list into some final type TM t (fromF), and also extract it back into a linked list (toF).

Constructors

Done (i a)

More (t f (F t i f) a)

injectF :: forall t f. Tensor t => f ~> F t (I t) f

If we have Tensor t, we can make a singleton F.

retractF :: forall t f. (Monoidal t, C (TM t) f) => F t (I t) f ~> f

If we have Monoidal t, we can collapse all t fs in the F into a single f.

interpretF :: forall t f g. (Monoidal t, C (TM t) g) => (f ~> g) -> F t (I t) f ~> g

If we have Monoidal t, we can interpret all of the fs in the F into a final target context g, given an f-to-g interpreting function.

newtype WrappedHBifunctor t (f :: Type -> Type) (g :: Type -> Type) a

Useful newtype to allow us to derive an HFunctor instance from any instance of HBifunctor, using -XDerivingVia.

For example, because we have instance HBifunctor Day, we can write:

deriving via (WrappedHBifunctor Day f) instance HFunctor (Day f)

to give us an automatic HFunctor instance and save us some work.

Constructors

WrappedHBifunctor

Fields

unwrapHBifunctor :: t f g a

data JoinT t f a

Constructors

JoinT

Fields

runJoinT :: t f f a