Data.HBifunctor.Tensor

Description

This module provides tools for working with binary functor combinators.

Data.Functor.HFunctor deals with single functor combinators (transforming a single functor). This module provides tools for working with combinators that combine and mix two functors "together".

The binary analog of HFunctor is HBifunctor: we can map a structure-transforming function over both of the transformed functors.

The binary analog of Interpret is Monoidal (and Tensor). If your combinator is an instance of Monoidal, it means that you can "squish" both arguments together into an Interpret. For example:

toMF :: (f :*: f) a -> ListF f a
toMF :: Comp f f a -> Free f a
toMF :: Day f f a -> Ap f a

Tensor

class Associative t => Tensor t where

A HBifunctor can be a Tensor if:

There is some identity i where t i f is equivalent to just f. That is, "enhancing" f with t i does nothing.

t is associative: f t (g t h) is equivalent to (f t g) t h).

The methods in this class provide us useful ways of navigating a Tensor t with respect to this property.

Realistically, there won't be any Tensor instances that are not also Monoidal instances. The two classes are separated only to help organize functionality into cleaner sub-divisions.

Associated Types

type I t :: Type -> Type

The identity of Tensor t. If you "combine" f with the identity, it leaves f unchanged.

For example, the identity of :*: is Proxy. This is because

(Proxy :*: f) a

is equivalent to just

f a

:*:-ing f with Proxy gives you no additional structure.

Methods

intro1 :: f ~> t f (I t)

intro2 :: g ~> t (I t) g

elim1 :: Functor f => t f (I t) ~> f

elim2 :: Functor g => t (I t) g ~> g

rightIdentity :: (Tensor t, Functor f) => f <~> t f (I t)

leftIdentity :: (Tensor t, Functor g) => g <~> t (I t) g

voidLeftIdentity :: f <~> (Void1 :+: f)

voidRightIdentity :: f <~> (Void1 :+: f)

Monoidal

class (Tensor t, Semigroupoidal t, Interpret (MF t)) => Monoidal t where

For some tensors t, you can represt the act of repeatedly combining the same functor an arbitrary amount of times:

t f f                    -- 2 times
t f (t f f)              -- 3 times
t f (t f (t f f))        -- 4 times
t f (t f (t f (t f f)))  -- 5 times

Sometimes, we have a type that can describe this repeated combination. For example, ListF f is the type that contains f :*:d with itself many number of times, and Ap is the type that contains f Dayd with itself many number of times.

ListF [x, y]       == x :*: y
ListF [x, y, z]    == x :*: y :*: z
ListF [x, y, z, q] == x :*: y :*: z :*: q

This is convenient because it allows you to represent repeated applications of t as a single data type.

For example, Day f f can be interpreted as "two sequenced fs", allowing you to specify "I want exactly two sequenced fs". If you want to specify "I want 0, 1, or many fs sequenced after each other", then you can use Ap f.

And, f :*: f can be interpreted as "a free selection of two fs", allowing you to specify "I have to fs that I can use". If you want to specify "I want 0, 1, or many different fs that I can use", you can use ListF f.

The Monoidal class unifies different such patterns. The associated type MF is the "repeated aplications of t" type.

See documentation of Data.Functor.Tensor for information on how to define instances of this typeclass.

Minimal complete definition

appendMF, splitSF, splittingMF

Associated Types

type MF t :: (Type -> Type) -> Type -> Type

Methods

appendMF :: t (MF t f) (MF t f) ~> MF t f

splitSF :: SF t f ~> t f (MF t f)

splittingMF :: MF t f <~> (I t :+: t f (MF t f))

toMF :: t f f ~> MF t f

fromSF :: SF t f ~> MF t f

pureT :: C (MF t) f => I t ~> f

nilMF :: forall t f. Monoidal t => I t ~> MF t f

consMF :: forall t f. Monoidal t => t f (MF t f) ~> MF t f

unconsMF :: forall t f. Monoidal t => MF t f ~> (I t :+: t f (MF t f))

unmatchMF :: forall t f. Monoidal t => (I t :+: SF t f) ~> MF t f

Utility

inL :: forall t f g a. (Monoidal t, C (MF t) g) => f a -> t f g a

Convenient wrapper over intro1 that lets us introduce an arbitrary functor g to the right of an f.

inR :: forall t f g a. (Monoidal t, C (MF t) f) => g a -> t f g a

Convenient wrapper over intro2 that lets us introduce an arbitrary functor f to the right of a g.

Chain

data Chain t i f a

A useful construction that works like a "linked list" of t f applied to itself multiple times. That is, it contains t f f, t f (t f f), t f (t f (t f f)), etc.

If t is Monoidal, then it means we can "collapse" this linked list into some final type MF t (fromF), and also extract it back into a linked list (toF).

Constructors

Done (i a)

More (t f (Chain t i f) a)

foldChain :: forall t i f g. HBifunctor t => (i ~> g) -> (t f g ~> g) -> Chain t i f ~> g

unfoldChain :: forall t f g i. HBifunctor t => (g ~> (i :+: t f g)) -> g ~> Chain t i f

unrollMF :: forall t f. Monoidal t => MF t f ~> Chain t (I t) f

rerollMF :: forall t f. Monoidal t => Chain t (I t) f ~> MF t f

unrollingMF :: forall t f. Monoidal t => MF t f <~> Chain t (I t) f

fromChain1 :: forall t f. Monoidal t => Chain1 t f ~> Chain t (I t) f