Data.HFunctor.Internal

Documentation

class HFunctor t where

An HFunctor can be thought of a unary "functor transformer" --- a basic functor combinator. It takes a functor as input and returns a functor as output.

It "enhances" a functor with extra structure (sort of like how a monad transformer enhances a Monad with extra structure).

As a uniform inteface, we can "swap the underlying functor" (also sometimes called "hoisting"). This is what hmap does: it lets us swap out the f in a t f for a t g.

For example, the free monad Free takes a Functor and returns a new Functor. In the process, it provides a monadic structure over f. hmap lets us turn a Free f into a Free g: a monad built over f can be turned into a monad built over g.

For the ability to move in and out of the enhanced functor, see Inject and Interpret.

Methods

hmap :: (f ~> g) -> t f ~> t g

If we can turn an f into a g, then we can turn a t f into a t g.

It must be the case that

hmap id == id

Essentially, t f adds some "extra structure" to f. hmap must swap out the functor, without affecting the added structure.

For example, ListF f a is essentially a list of f as. If we hmap to swap out the f as for g as, then we must ensure that the "added structure" (here, the number of items in the list, and the ordering of those items) remains the same. So, hmap must preserve the number of items in the list, and must maintain the ordering.

The law hmap id == id is a way of formalizing this property.

class HBifunctor t where

A HBifunctor is like an HFunctor, but it enhances two different functors instead of just one.

Usually, it enhaces them "together" in some sort of combining way.

This typeclass provides a uniform instance for "swapping out" or "hoisting" the enhanced functors. We can hoist the first one with hleft, the second one with hright, or both at the same time with hbimap.

For example, the f :*: g type gives us "both f and g":

data (f :*: g) a = f a :*: g a

It combines both f and g into a unified structure --- here, it does it by providing both f and g.

The single law is:

hbimap id id == id

This ensures that hleft, hright, and hbimap do not affect the structure that t adds on top of the underlying functors.

Minimal complete definition

hleft, hright | hbimap

Methods

hleft :: (f ~> j) -> t f g ~> t j g

Swap out the first transformed functor.

hright :: (g ~> k) -> t f g ~> t f k

Swap out the second transformed functor.

hbimap :: (f ~> j) -> (g ~> k) -> t f g ~> t j k

Swap out both transformed functors at the same time.

newtype WrappedHBifunctor t (f :: Type -> Type) (g :: Type -> Type) a

Useful newtype to allow us to derive an HFunctor instance from any instance of HBifunctor, using -XDerivingVia.

For example, because we have instance HBifunctor Day, we can write:

deriving via (WrappedHBifunctor Day f) instance HFunctor (Day f)

to give us an automatic HFunctor instance and save us some work.

Constructors

WrapHBifunctor

Fields

unwrapHBifunctor :: t f g a

sumSum :: (f :+: g) <~> Sum f g

prodProd :: (f :*: g) <~> Product f g

unsafePlus :: forall f proxy r. Alternative f => proxy f -> (Plus f => r) -> r

For any Alternative f, produce a value that would require Plus f.

Always use with concrete and specific f only, and never use with any f that already has a Plus instance.

The Proxy argument allows you to specify which specific f you want to enhance. You can pass in something like Proxy @MyFunctor.

unsafeApply :: forall f proxy r. Applicative f => proxy f -> (Apply f => r) -> r

For any Applicative f, produce a value that would require Apply f.

Always use with concrete and specific f only, and never use with any f that already has a Apply instance.

The Proxy argument allows you to specify which specific f you want to enhance. You can pass in something like Proxy @MyFunctor.

unsafeBind :: forall f proxy r. Monad f => proxy f -> (Bind f => r) -> r

For any Monad f, produce a value that would require Bind f.

Always use with concrete and specific f only, and never use with any f that already has a Bind instance.

The Proxy argument allows you to specify which specific f you want to enhance. You can pass in something like Proxy @MyFunctor.