Control.Applicative.Step

Fixed Points

data Step f a

The fixed point of applications of :+: (functor sums).

Intuitively, in an infinite f :+: f :+: f :+: f ..., you have exactly one f somewhere. A Step f a has that f, with a Natural giving you "where" the f is in the long chain.

interpreting it requires no constraint on the target context.

Constructors

Step

Fields

stepPos :: Natural

stepVal :: f a

newtype Steps f a

The fixed point of applications of TheseT.

Intuitively, in an infinite f TheseT f TheseT f TheseT f ..., each of those infinite positions may have an f in them. However, because of the at-least-one nature of TheseT, we know we have at least one f at one position somewhere.

A Steps f a has potentially many fs, each stored at a different Natural position, with the guaruntee that at least one f exists.

interpreting it requires at least an Alt instance in the target context, since we have to handle potentially more than one f. However, we don't fully need Plus, since we know we always have at least one f.

Constructors

Steps

Fields

getSteps :: NEMap Natural (f a)

Void

data VoidT a

The identity functor of :+: (and also TheseT)

absurdT :: VoidT a -> f a

We have a natural transformation between VoidT and any other functor f with no constraints.