Control.Applicative.Step

Description

This module provides functor combinators that are the fixed points of applications of :+: and These1. They are useful for their Interpret instances, along with their relationship to the Monoidal instances of :+: and These1.

Fixed Points

data Step f a

An f a, along with a Natural index.

Step f a ~ (Natural, f a)
Step f   ~ ((,) Natural) :.: f       -- functor composition

It is the fixed point of infinite applications of :+: (functor sums).

Intuitively, in an infinite f :+: f :+: f :+: f ..., you have exactly one f somewhere. A Step f a has that f, with a Natural giving you "where" the f is in the long chain.

Can be useful for using with the Monoidal instance of :+:.

interpreting it requires no constraint on the target context.

Note that this type and its instances equivalent to EnvT (Sum Natural).

Constructors

Step

Fields

stepPos :: Natural

stepVal :: f a

newtype Steps f a

A non-empty map of Natural to f a. Basically, contains multiple f as, each at a given Natural index.

Steps f a ~ Map Natural (f a)
Steps f   ~ Map Natural :.: f       -- functor composition

It is the fixed point of applications of TheseT.

You can think of this as an infinite sparse array of f as.

Intuitively, in an infinite f `TheseT` f `TheseT` f `TheseT` f ..., each of those infinite positions may have an f in them. However, because of the at-least-one nature of TheseT, we know we have at least one f at one position somewhere.

A Steps f a has potentially many fs, each stored at a different Natural position, with the guaruntee that at least one f exists.

Can be useful for using with the Monoidal instance of TheseT.

interpreting it requires at least an Alt instance in the target context, since we have to handle potentially more than one f.

Constructors

Steps

Fields

getSteps :: NEMap Natural (f a)

data Flagged f a

An f a, along with a Bool flag

Flagged f a ~ (Bool, f a)
Flagged f   ~ ((,) Bool) :.: f       -- functor composition

Creation with inject or pure uses False as the boolean.

You can think of it as an f a that is "flagged" with a boolean value, and that value can indicuate whether or not it is "pure" (made with inject or pure) as False, or "impure" (made from some other source) as True. However, False may be always created directly, of course, using the constructor.

You can think of it like a Step that is either 0 or 1, as well.

interpreting it requires no constraint on the target context.

This type is equivalent (along with its instances) to:

HLift IdentityT

EnvT Any

Constructors

Flagged

Fields

flaggedFlag :: Bool

flaggedVal :: f a

Steppers

stepUp :: (f :+: Step f) ~> Step f

Unshift an item into a Step. Because a Step f is f :+: f :+: f :+: f :+: ... forever, this basically conses an additional possibility of f to the beginning of it all.

You can think of it as reassociating

f :+: ( f :+: f :+: f :+: ...)

into

f :+: f :+: f :+: f :+: ...

stepUp (L1 "hello")
-- Step 0 "hello"
stepUp (R1 (Step 1 "hello"))
-- Step 2 "hello"

Forms an isomorphism with stepDown (see stepping).

stepDown :: Step f ~> (f :+: Step f)

Pop off the first item in a Step. Because a Step f is f :+: f :+: f :+: ... forever, this matches on the first branch.

You can think of it as reassociating

f :+: f :+: f :+: f :+: ...

into

f :+: ( f :+: f :+: f :+: ...)

stepDown (Step 2 "hello")
-- R1 (Step 1 "hello")
stepDown (Step 0 "hello")
-- L1 "hello"

Forms an isomorphism with stepUp (see stepping).

stepping :: Step f <~> (f :+: Step f)

"Uncons and cons" an f branch before a Step. This is basically a witness that stepDown and stepUp form an isomorphism.

stepsUp :: These1 f (Steps f) ~> Steps f

Unshift an item into a Steps. Because a Steps f is f These1 f These1 f These1 f These1 ... forever, this basically conses an additional possibility of f to the beginning of it all.

You can think of it as reassociating

f These1 ( f These1 f These1 f These1 ...)

into

f These1 f These1 f These1 f These1 ...

If you give:

This1, then it returns a singleton Steps with one item at index 0

That1, then it shifts every item in the given Steps up one index.

These1, then it shifts every item in the given Steps up one index, and adds the given item (the f) at index zero.

Forms an isomorphism with stepDown (see stepping).

stepsDown :: Steps f ~> These1 f (Steps f)

Pop off the first item in a Steps. Because a Steps f is f These1 f These1 f These1 ... forever, this matches on the first branch.

You can think of it as reassociating

f These1 f These1 f These1 f These1 ...

into

f These1 ( f These1 f These1 f These1 ...)

It returns:

This1 if the first item is the only item in the Steps

That1 if the first item in the Steps is empty, but there are more items left. The extra items are all shfited down.

These1 if the first item in the Steps exists, and there are also more items left. The extra items are all shifted down.

Forms an isomorphism with stepsUp (see steppings).

steppings :: Steps f <~> These1 f (Steps f)

"Uncons and cons" an f branch before a Steps. This is basically a witness that stepsDown and stepsUp form an isomorphism.

Void

absurd1 :: V1 a -> f a

We have a natural transformation between V1 and any other functor f with no constraints.

data Void2 a b

Void2 a b is uninhabited for all a and b.

absurd2 :: Void2 f a -> t f a

If you treat a Void2 f a as a functor combinator, then absurd2 lets you convert from a Void2 f a into a t f a for any functor combinator t.

data Void3 a b c

Void3 a b is uninhabited for all a and b.

absurd3 :: Void3 f g a -> t f g a

If you treat a Void3 f a as a binary functor combinator, then absurd3 lets you convert from a Void3 f a into a t f a for any functor combinator t.