Control.Applicative.Step

Description

This module provides functor combinators that are the fixed points of applications of :+: and TheseT. They are useful for their Interpret instances, along with their relationship to the Monoidal insatnces of :+: and TheseT.

Fixed Points

data Step f a

An f a, along with a Natural index.

Step f a ~ (Natural, f a)
Step f   ~ ((,) Natural) :.: f       -- functor composition

It is the fixed point of infinite applications of :+: (functor sums).

Intuitively, in an infinite f :+: f :+: f :+: f ..., you have exactly one f somewhere. A Step f a has that f, with a Natural giving you "where" the f is in the long chain.

Can be useful for using with the Monoidal instance of :+:.

interpreting it requires no constraint on the target context.

Constructors

Step

Fields

stepPos :: Natural

stepVal :: f a

newtype Steps f a

A non-empty map of Natural to f a. Basically, contains multiple f as, each at a given Natural index.

Steps f a ~ Map Natural (f a)
Steps f   ~ Map Natural :.: f       -- functor composition

It is the fixed point of applications of TheseT.

Intuitively, in an infinite f `TheseT` f `TheseT` f `TheseT` f ..., each of those infinite positions may have an f in them. However, because of the at-least-one nature of TheseT, we know we have at least one f at one position somewhere.

A Steps f a has potentially many fs, each stored at a different Natural position, with the guaruntee that at least one f exists.

Can be useful for using with the Monoidal instance of TheseT.

interpreting it requires at least an Alt instance in the target context, since we have to handle potentially more than one f.

Constructors

Steps

Fields

getSteps :: NEMap Natural (f a)

stepUp :: (f :+: Step f) ~> Step f

Unshift an item into a Step. Because a Step f is f :+: f :+: f :+: f :+: ... forever, this basically conses an additional possibility of f to the beginning of it all.

You can think of it as reassociating

f :+: ( f :+: f :+: f :+: ...)

into

f :+: f :+: f :+: f :+: ...

Forms an isomorphism with stepDown (see stepping).

stepDown :: Step f ~> (f :+: Step f)

Pop off the first iteem in a Step. Because a Step f is f :+: f :+: f :+: ... forever, this matches on the first branch.

You can think of it as reassociating

f :+: f :+: f :+: f :+: ...

into

f :+: ( f :+: f :+: f :+: ...)

Forms an isomorphism with stepUp (see stepping).

stepping :: Step f <~> (f :+: Step f)

"Uncons and cons" an f branch before a Step. This is basically a witness that stepDown and stepUp form an isomorphism.

Void

data Void1 a

Void1 is a functor that is uninhabited for all inputs. That is, Void1 a is uninhabited for all a.

The identity functor of :+: (and also TheseT)

absurd1 :: Void1 a -> f a

We have a natural transformation between Void1 and any other functor f with no constraints.

data Void2 a b

Void2 a b is uninhabited for all a and b.

absurd2 :: Void2 f a -> t f a

If you treat a Void2 f a as a functor combinator, then absurd2 lets you convert from a Void2 f a into a t f a for any functor combinator t.