Control.Monad.Freer.Church

Description

The church-encoded Freer Monad. Basically provides the free monad in a way that is compatible with HFunctor and Interpret, and :.: or Compose in a way that is compatible with HBifunctor and Tensor and Monoidal.

Documentation

newtype Free f a

A Free f is f enhanced with "sequential binding" capabilities. It allows you to sequence multiple fs one after the other, and also to determine "what f to sequence" based on the result of the computation so far.

Essentially, you can think of this as "giving f a Monad instance", with all that that entails (return, >>=, etc.).

Lift f into it with inject :: f a -> Free f a. When you finally want to "use" it, you can interpret it into any monadic context:

interpret
    :: Monad g
    => (forall x. f x -> g x)
    -> Free f a
    -> g a

Under the hood, this is the Church-encoded Freer monad. It's Free, or F, but in a way that is compatible with HFunctor and Interpret.

Constructors

Free

Fields

runFree :: forall r. (a -> r) -> (forall s. f s -> (s -> r) -> r) -> r

reFree :: (MonadFree f m, Functor f) => Free f a -> m a

Convert a Free f into any instance of MonadFree f.

data Comp f g a

Functor composition. 'type Comp' f g a is equivalent to f (g a), and the 'pattern Comp' pattern synonym is a way of getting the f (g a) in a 'type Comp' f g a.

For example, Maybe (IO Bool) is 'type Comp' Maybe IO Bool.

This is mostly useful for its typeclass instances: in particular, Functor, Applicative, HBifunctor, and Monoidal.

This is essentially a version of :.: and Compose that allows for an HBifunctor instance.

It is slightly less performant. Using comp . unComp every once in a while will concretize a 'type Comp' value (if you have Functor f) and remove some indirection if you have a lot of chained operations.

The "free monoid" over 'type Comp' is Free.

Constructors

(f x) :>>= (x -> g a)

Bundled Patterns

pattern Comp :: Functor f => f (g a) -> Comp f g a
Pattern match on and construct a 'type Comp' f g a as if it were f (g a).

comp :: f (g a) -> Comp f g a

"Smart constructor" for 'type Comp' that doesn't require Functor f.