Data.Functor.Apply.Free

Description

The free Apply. Provides Ap1 and various utility methods. See Ap1 for more details.

Ideally Ap1 would be in the free package. However, it is defined here for now.

Documentation

data Ap1 :: (Type -> Type) -> Type -> Type where

One or more fs convolved with itself.

Essentially:

Ap1 f
    ~ f                            -- one f
  :+: (f `Day' f)          -- two f's
  :+: (f `Day` f `Day` f)           -- three f's
  :+: (f `Day` f `Day` f `Day` f)  -- four f's
  :+: ...                          -- etc.

Useful if you want to promote an f to a situation with "at least one f sequenced with itself".

Mostly useful for its HFunctor and Interpret instance, along with its relationship with Ap and Day.

This is the free Apply --- Basically a "non-empty" Ap.

The construction here is based on Ap, similar to now NonEmpty is built on list.

Constructors

Ap1 :: f a -> Ap f (a -> b) -> Ap1 f b

Bundled Patterns

pattern DayAp1 :: Day f (Ap f) a -> Ap1 f a
An Ap1 f is just a Day f (Ap f). This bidirectional pattern synonym lets you treat it as such.

toAp :: Ap1 f ~> Ap f

An Ap1 is a "non-empty" Ap; this function "forgets" the non-empty property and turns it back into a normal Ap.

fromAp :: Ap f ~> (Identity :+: Ap1 f)

Convert an Ap into an Ap1 if possible. If the Ap was "empty", return the Pure value instead.

liftAp1 :: f ~> Ap1 f

Embed an f into Ap1.

retractAp1 :: Apply f => Ap1 f ~> f

Extract the f out of the Ap1.

retractAp1 . liftAp1 == id

runAp1 :: Apply g => (f ~> g) -> Ap1 f ~> g

Interpret an Ap f into some Apply context g.