Data.Functor.Combinator

Single Functors

type (~>) (f :: k -> Type) (g :: k -> Type) = forall (x :: k). f x -> g x infixr 0

A natural transformation from f to g.

class HFunctor t where

An HFunctor can be thought of a unary "functor transformer" --- a basic functor combinator. It takes a functor as input and returns a functor as output.

It "enhances" a functor with extra structure (sort of like how a monad transformer enhances a Monad with extra structure).

As a uniform inteface, we can "swap the underlying functor" (also sometimes called "hoisting"). This is what hmap does: it lets us swap out the f in a t f for a t g.

For example, the free monad Free takes a Functor and returns a new Functor. In the process, it provides a monadic structure over f. hmap lets us turn a Free f into a Free g: a monad built over f can be turned into a monad built over g.

For the ability to move in and out of the enhanced functor, see Interpret.

Methods

hmap :: (f ~> g) -> t f ~> t g

If we can turn an f into a g, then we can turn a t f into a t g.

It must be the case that

hmap id == id

Essentially, t f adds some "extra structure" to f. hmap must swap out the functor, without affecting the added structure.

For example, ListF f a is essentially a list of f as. If we hmap to swap out the f as for g as, then we must ensure that the "added structure" (here, the number of items in the list) remains the same. So, hmap must preserve the number of items in the list.

The law hmap id == id is a way of formalizing this informal property.

class HFunctor t => Interpret t where

An Interpret lets us move in and out of the "enhanced" Functor.

For example, Free f is f enhanced with monadic structure. We get:

inject    :: f a -> Free f a
interpret :: Monad m => (forall x. f x -> m x) -> Free f a -> m a

inject will let us use our f inside the enhanced Free f. interpret will let us "extract" the f from a Free f if we can give an interpreting function that interprets f into some target Monad.

The type family C tells us the typeclass constraint of the "target" functor. For Free, it is Monad, but for other Interpret instances, we might have other constraints.

We enforce that:

interpret id . inject == id

That is, if we lift a value into our structure, then immediately interpret it out as itself, it should lave the value unchanged.

Minimal complete definition

inject, (retract | interpret)

Associated Types

type C t :: (Type -> Type) -> Constraint

The constraint on the target context of interpret.

Methods

inject :: f ~> t f

Lift an f into the enhanced t f structure. Analogous to lift from MonadTrans.

retract :: C t f => t f ~> f

Remove the f out of the enhanced t f structure, provided that f satisfies the necessary constraints. If it doesn't, it needs to be properly interpreted out.

interpret :: C t g => (f ~> g) -> t f ~> g

Given an "interpeting function" from f to g, interpret the f out of the t f into a final context g.

extractI :: (Interpret t, C t f, Copointed f) => t f a -> a

Useful wrapper over retract to allow you to directly extract an a from a t f a, if f is a valid retraction from t, and f is an instance of Copointed.

Useful fs include Identity or related newtype wrappers from base:

extractI
    :: (Interpret t, C t Identity)
    => t Identity a
    -> a

getI :: (Interpret t, C t (Const b)) => (forall x. f x -> b) -> t f a -> b

Useful wrapper over interpret to allow you to directly extract a value b out of the t f a, if you can convert f x into b.

Note that depending on the constraints on the interpretation of t, you may have extra constraints on b.

If C t is Unconstrained, there are no constraints on b

If C t is Apply, b needs to be an instance of Semigroup

If C t is Applicative, b needs to be an instance of Monoid

For some constraints (like Monad), this will not be usable.

-- get the length of the Map String in the Step.
collectI length
     :: Step (Map String) Bool
     -> [Int]

collectI :: (Interpret t, C t (Const [b])) => (forall x. f x -> b) -> t f a -> [b]

Useful wrapper over getI to allow you to collect a b from all instances of f inside a t f a.

This will work if C t is Unconstrained, Apply, or Applicative.

-- get the lengths of all Map Strings in the Ap.
collectI length
     :: Ap (Map String) Bool
     -> [Int]

Multi-Functors

class HBifunctor t where

A HBifunctor is like an HFunctor, but it enhances two different functors instead of just one.

Usually, it enhaces them "together" in some sort of combining way.

This typeclass provides a uniform instance for "swapping out" or "hoisting" the enhanced functors. We can hoist the first one with hleft, the second one with hright, or both at the same time with hbimap.

For example, the f :*: g type gives us "both f and g":

data (f :*: g) a = f a :*: g a

It combines both f and g into a unified structure --- here, it does it by providing both f and g.

The single law is:

hbimap id id == id

This ensures that hleft, hright, and hbimap do not affect the structure that t adds on top of the underlying functors.

Minimal complete definition

hleft, hright | hbimap

Methods

hleft :: (f ~> j) -> t f g ~> t j g

Swap out the first transformed functor.

hright :: (g ~> k) -> t f g ~> t f k

Swap out the second transformed functor.

hbimap :: (f ~> j) -> (g ~> k) -> t f g ~> t j k

Swap out both transformed functors at the same time.

class HBifunctor t => Tensor t

A HBifunctor can be a Tensor if:

There is some identity i where t i f is equivalent to just f. That is, "enhancing" f with t i does nothing.

t is associative: f t (g t h) is equivalent to (f t g) t h).

The methods in this class provide us useful ways of navigating a Tensor t with respect to this property.

Minimal complete definition

intro1, intro2, elim1, elim2, assoc, disassoc

Associated Types

type I t :: Type -> Type

The identity of Tensor t. If you "combine" f with the identity, it leaves f unchanged.

For example, the identity of :*: is Proxy. This is because

(Proxy :*: f) a

is equivalent to just

f a

:*:-ing f with Proxy gives you no additional structure.

class (Tensor t, Interpret (TM t)) => Monoidal t where

For some tensors t, you can represt the act of repeatedly combining the same functor an arbitrary amount of times:

t f f                    -- 2 times
t f (t f f)              -- 3 times
t f (t f (t f f))        -- 4 times
t f (t f (t f (t f f)))  -- 5 times

Sometimes, we have a type that can describe this repeated combination. For example, ListF f is the type that contains f :*:d with itself many number of times, and Ap is the type that contains f Dayd with itself many number of times.

ListF [x, y]       == x :*: y
ListF [x, y, z]    == x :*: y :*: z
ListF [x, y, z, q] == x :*: y :*: z :*: q

This is convenient because it allows you to represent repeated applications of t as a single data type.

For example, Day f f can be interpreted as "two sequenced fs", allowing you to specify "I want exactly two sequenced fs". If you want to specify "I want 0, 1, or many fs sequenced after each other", then you can use Ap f.

And, f :*: f can be interpreted as "a free selection of two fs", allowing you to specify "I have to fs that I can use". If you want to specify "I want 0, 1, or many different fs that I can use", you can use ListF f.

The Monoidal class unifies different such patterns. The associated type TM is the "repeated aplications of t" type.

See documentation of Data.Functor.Tensor for information on how to define instances of this typeclass.

Minimal complete definition

nilTM, consTM, unconsTM, appendTM | fromF, toF, appendF

Associated Types

type TM t :: (Type -> Type) -> Type -> Type

Methods

retractT :: C (TM t) f => t f f ~> f

A version of retract that works for a Tensor. It retracts both fs into a single f.

interpretT :: C (TM t) h => (f ~> h) -> (g ~> h) -> t f g ~> h

A version of interpret that works for a Tensor. It takes two interpreting functions, and interprets both joined functors one after the other into h.

pureT :: C (TM t) f => I t ~> f

If we have an instance of t, we can generate an f based on how it interacts with t.

Specialized (and simplified), this type is:

pureT @Day   :: Applicative f => a -> f a  -- pure
pureT @Comp  :: Monad f => a -> f a    -- return
pureT @(:*:) :: Plus f => f a          -- zero

toTM :: t f f ~> TM t f

Embed a direct application of f to itself into a TM t f.

inL :: forall t f g a. (Monoidal t, C (TM t) g) => f a -> t f g a

Convenient wrapper over intro1 that lets us introduce an arbitrary functor g to the right of an f.

inR :: forall t f g a. (Monoidal t, C (TM t) f, Functor g) => g a -> t f g a

Convenient wrapper over intro2 that lets us introduce an arbitrary functor f to the right of a g.

(!$!) :: (Monoidal t, C (TM t) (Const b)) => (forall x. f x -> b) -> (forall x. g x -> b) -> t f g a -> b infixr 5

Infix alias for getT

extractT :: (Monoidal t, C (TM t) f, Copointed f) => t f f a -> a

Useful wrapper over retractT to allow you to directly extract an a from a t f f a, if f is a valid retraction from t, and f is an instance of Copointed.

Useful fs include Identity or related newtype wrappers from base:

extractT
    :: (Monoidal t, C (TM t) Identity)
    => t Identity Identity a
    -> a

getT :: (Monoidal t, C (TM t) (Const b)) => (forall x. f x -> b) -> (forall x. g x -> b) -> t f g a -> b

Useful wrapper over interpret to allow you to directly extract a value b out of the t f a, if you can convert f x into b.

Note that depending on the constraints on the interpretation of t, you may have extra constraints on b.

If C (TM t) is Unconstrained, there are no constraints on b

If C (TM t) is Apply, b needs to be an instance of Semigroup

If C (TM t) is Applicative, b needs to be an instance of Monoid

For some constraints (like Monad), this will not be usable.

-- Return the length of either the list, or the Map, depending on which
--   one s in the +
length !*! length
    :: ([] :+: Map Int) Char
    -> Int

-- Return the length of both the list and the map, added together
(Sum . length) !*! (Sum . length)
    :: Day [] (Map Int) Char
    -> Sum Int

(!*!) :: (Monoidal t, C (TM t) h) => (f ~> h) -> (g ~> h) -> t f g ~> h infixr 5

Infix alias for interpretT

collectT :: (Monoidal t, C (TM t) (Const [b])) => (forall x. f x -> b) -> (forall x. g x -> b) -> t f g a -> [b]

Useful wrapper over getT to allow you to collect a b from all instances of f and g inside a t f g a.

This will work if C t is Unconstrained, Apply, or Applicative.

Combinators

data Coyoneda (f :: Type -> Type) a where

A covariant Functor suitable for Yoneda reduction

Constructors

Coyoneda :: forall (f :: Type -> Type) a b. (b -> a) -> f b -> Coyoneda f a

newtype ListF f a

A list of f as.

This is the Free Plus.

Constructors

ListF

Fields

runListF :: [f a]

data Ap (f :: Type -> Type) a

The free Applicative for a Functor f.

data Ap1 :: (Type -> Type) -> Type -> Type

The free Apply. Basically a "non-empty" Ap.

The construction here is based on Ap, similar to now NonEmpty is built on list.

data Alt (f :: Type -> Type) a

data Free f a

Church-encoded Freer monad

data Step f a

The fixed point of applications of :+: (functor sums).

Intuitively, in an infinite f :+: f :+: f :+: f ..., you have exactly one f somewhere. A Step f a has that f, with a Natural giving you "where" the f is in the long chain.

interpreting it requires no constraint on the target context.

Constructors

Step

Fields

stepPos :: Natural

stepVal :: f a

newtype Steps f a

The fixed point of applications of TheseT.

Intuitively, in an infinite f TheseT f TheseT f TheseT f ..., each of those infinite positions may have an f in them. However, because of the at-least-one nature of TheseT, we know we have at least one f at one position somewhere.

A Steps f a has potentially many fs, each stored at a different Natural position, with the guaruntee that at least one f exists.

interpreting it requires at least an Alt instance in the target context, since we have to handle potentially more than one f. However, we don't fully need Plus, since we know we always have at least one f.

Constructors

Steps

Fields

getSteps :: NEMap Natural (f a)

data Day (f :: Type -> Type) (g :: Type -> Type) a where

The Day convolution of two covariant functors.

Constructors

Day :: forall (f :: Type -> Type) (g :: Type -> Type) a b c. f b -> g c -> (b -> c -> a) -> Day f g a

data ((f :: k -> Type) :*: (g :: k -> Type)) (p :: k) :: forall k. (k -> Type) -> (k -> Type) -> k -> Type infixr 6

Products: encode multiple arguments to constructors

Constructors

(f p) :*: (g p) infixr 6

data ((f :: k -> Type) :+: (g :: k -> Type)) (p :: k) :: forall k. (k -> Type) -> (k -> Type) -> k -> Type infixr 5

Sums: encode choice between constructors

Constructors

L1 (f p)

R1 (g p)

data These1 f g a

A These f g a has either an f a, a g a, or both.

This is re-defined here from the these package, to avoid the high dependency footprint.

Constructors

This1 (f a)

That1 (g a)

These1 (f a) (g a)

type Comp = Cons Star

A version of :.: that permits HBifunctor, Tensor, and Monoidal instances.

newtype Final c f a

A simple way to inject/reject into any eventual typeclass.

In a way, this is the "ultimate" Interpret instance. You can use this to inject an f into a free structure of any typeclass. If you want f to have a Monad instance, for example, just use

inject :: f a -> Final Monad f a

When you want to eventually interpret out the data, use:

interpret :: (f ~> g) -> Final c f a -> g a

Essentially, Final c is the "free c". Final Monad is the free Monad, etc.

Note that this doesn't have instances for all the typeclasses you could lift things into; you probably have to define your own if you want to use Final c as an instance of c (using liftFinal0, liftFinal1, liftFinal2 for help).

Constructors

Final

Fields

runFinal :: forall g. c g => (forall x. f x -> g x) -> g a

class Interpret t => FreeOf c t | t -> c where

A typeclass associating a free structure with the typeclass it is free on.

This essentially lists instances of Interpret where a "trip" through Final will leave it unchanged.

fromFree . toFree == id
toFree . fromFree == id

This can be useful because Final doesn't have a concrete structure that you can pattern match on and inspect, but t might. This lets you work on a concrete structure if you desire.

Minimal complete definition

Nothing

Methods

fromFree :: t f ~> Final c f

toFree :: Functor f => Final c f ~> t f

fromFree :: C t (Final c f) => t f ~> Final c f

toFree :: c (t f) => Final c f ~> t f

data Cons :: ((Type -> Type) -> Type -> Type -> Type) -> (Type -> Type) -> (Type -> Type) -> Type -> Type where

A lot of tensors can be factored out into a Cons on some parameterized profunctor. For example, we have:

:.: = Cons Star
Day = Cons Static
:*: = Cons ProdK

In the case of :.:, this actually allows us to have HFunctor and Interpret instances for functor composition, because the representation of :.: normally does not allow this. This gives us a proper tensor to form Free, which makes the free monad work nicely within our system.

In the normal case, you can just use Day and :*: instead of DayCons = Cons Static and ProdCons = Cons ProdK. The only one that gives us extra power is Comp = Cons Star.

You can use DayCons and ProdCons as Day and :*: using the provided pattern synonyms.

Bundled Patterns

pattern Comp :: Functor f => f (g a) -> Comp f g a

pattern DayCons :: Functor g => Day f g a -> DayCons f g a

pattern ProdCons :: Functor f => (f :*: g) a -> ProdCons f g a

class Monoidal (Cons p) => Uncons p t | t -> p where

Associates each Tensor t with its Cons equivalent, Cons p.

t need not always be a legal Tensor.

Methods

uncons :: Functor g => t f g ~> Cons p f g

cons :: Functor f => Cons p f g ~> t f g