Data.Functor.HFunctor

Documentation

class HFunctor t where

An HFunctor can be thought of a unary "functor transformer" --- a basic functor combinator. It takes a functor as input and returns a functor as output.

It "enhances" a functor with extra structure (sort of like how a monad transformer enhances a Monad with extra structure).

As a uniform inteface, we can "swap the underlying functor" (also sometimes called "hoisting"). This is what hmap does: it lets us swap out the f in a t f for a t g.

For example, the free monad Free takes a Functor and returns a new Functor. In the process, it provides a monadic structure over f. hmap lets us turn a Free f into a Free g: a monad built over f can be turned into a monad built over g.

For the ability to move in and out of the enhanced functor, see Interpret.

Methods

hmap :: (f ~> g) -> t f ~> t g

If we can turn an f into a g, then we can turn a t f into a t g.

It must be the case that

hmap id == id

Essentially, t f adds some "extra structure" to f. hmap must swap out the functor, without affecting the added structure.

For example, ListF f a is essentially a list of f as. If we hmap to swap out the f as for g as, then we must ensure that the "added structure" (here, the number of items in the list) remains the same. So, hmap must preserve the number of items in the list.

The law hmap id == id is a way of formalizing this informal property.

class HFunctor t => Interpret t where

An Interpret lets us move in and out of the "enhanced" Functor.

For example, Free f is f enhanced with monadic structure. We get:

inject    :: f a -> Free f a
interpret :: Monad m => (forall x. f x -> m x) -> Free f a -> m a

inject will let us use our f inside the enhanced Free f. interpret will let us "extract" the f from a Free f if we can give an interpreting function that interprets f into some target Monad.

The type family C tells us the typeclass constraint of the "target" functor. For Free, it is Monad, but for other Interpret instances, we might have other constraints.

We enforce that:

interpret id . inject == id

That is, if we lift a value into our structure, then immediately interpret it out as itself, it should lave the value unchanged.

Minimal complete definition

inject, (retract | interpret)

Associated Types

type C t :: (Type -> Type) -> Constraint

The constraint on the target context of interpret.

Methods

inject :: f ~> t f

Lift an f into the enhanced t f structure. Analogous to lift from MonadTrans.

retract :: C t f => t f ~> f

Remove the f out of the enhanced t f structure, provided that f satisfies the necessary constraints. If it doesn't, it needs to be properly interpreted out.

interpret :: C t g => (f ~> g) -> t f ~> g

Given an "interpeting function" from f to g, interpret the f out of the t f into a final context g.

extractI :: (Interpret t, C t f, Copointed f) => t f a -> a

Useful wrapper over retract to allow you to directly extract an a from a t f a, if f is a valid retraction from t, and f is an instance of Copointed.

Useful fs include Identity or related newtype wrappers from base:

extractI
    :: (Interpret t, C t Identity)
    => t Identity a
    -> a

getI :: (Interpret t, C t (Const b)) => (forall x. f x -> b) -> t f a -> b

Useful wrapper over interpret to allow you to directly extract a value b out of the t f a, if you can convert f x into b.

Note that depending on the constraints on the interpretation of t, you may have extra constraints on b.

If C t is Unconstrained, there are no constraints on b

If C t is Apply, b needs to be an instance of Semigroup

If C t is Applicative, b needs to be an instance of Monoid

For some constraints (like Monad), this will not be usable.

-- get the length of the Map String in the Step.
collectI length
     :: Step (Map String) Bool
     -> [Int]

collectI :: (Interpret t, C t (Const [b])) => (forall x. f x -> b) -> t f a -> [b]

Useful wrapper over getI to allow you to collect a b from all instances of f inside a t f a.

This will work if C t is Unconstrained, Apply, or Applicative.

-- get the lengths of all Map Strings in the Ap.
collectI length
     :: Ap (Map String) Bool
     -> [Int]

class (c a, d a) => AndC c d a