Data.Functor.HFunctor.Final

Documentation

newtype Final c f a

A simple way to inject/reject into any eventual typeclass. Essentially, Final c is the "free c". Final Monad is the free Monad, etc.

Useful for lifting a functor f into arbitrary structure given by a typeclass.

Note that this doesn't have instances for all the typeclasses; you probably have to define your own if you want to use Final c as an instance of c (using liftFinal0, liftFinal1, liftFinal2 for help). This is mostly meant to be usable as a final Interpret, with

inject    :: f a -> Final c f a
interpret :: (f ~> g) -> Final c f a -> g a

Constructors

Final

Fields

runFinal :: forall g. c g => (forall x. f x -> g x) -> g a

fromFinal :: (Interpret t, c (t f)) => Final c f ~> t f

Concretize a Final.

toFinal :: (Interpret t, C t (Final c f)) => t f ~> Final c f

Finalize an Interpret instance.

toFinal :: Coyoneda f ~> Final Functor f
toFinal :: Ap f ~> Final Applicative f
toFinal :: Alt f ~> Final Alternative f
toFinal :: Free f ~> Final Monad f

Note that the instance of c for Final c must be defined.

Should form an isomorphism with fromFinal

hoistFinalC :: (forall g x. (c g => g x) -> d g => g x) -> Final c f a -> Final d f a

liftFinal0 :: (forall g. c g => g a) -> Final c f a

liftFinal1 :: (forall g. c g => g a -> g b) -> Final c f a -> Final c f b

liftFinal2 :: (forall g. c g => g a -> g b -> g d) -> Final c f a -> Final c f b -> Final c f d