Data.HBifunctor

Description

This module provides an abstraction for "two-argument functor combinators", HBifunctor, as well as some useful combinators.

Documentation

class HBifunctor t where

A HBifunctor is like an HFunctor, but it enhances two different functors instead of just one.

Usually, it enhaces them "together" in some sort of combining way.

This typeclass provides a uniform instance for "swapping out" or "hoisting" the enhanced functors. We can hoist the first one with hleft, the second one with hright, or both at the same time with hbimap.

For example, the f :*: g type gives us "both f and g":

data (f :*: g) a = f a :*: g a

It combines both f and g into a unified structure --- here, it does it by providing both f and g.

The single law is:

hbimap id id == id

This ensures that hleft, hright, and hbimap do not affect the structure that t adds on top of the underlying functors.

Minimal complete definition

hleft, hright | hbimap

Methods

hleft :: (f ~> j) -> t f g ~> t j g

Swap out the first transformed functor.

hright :: (g ~> k) -> t f g ~> t f k

Swap out the second transformed functor.

hbimap :: (f ~> j) -> (g ~> k) -> t f g ~> t j k

Swap out both transformed functors at the same time.

newtype WrappedHBifunctor t (f :: Type -> Type) (g :: Type -> Type) a

Useful newtype to allow us to derive an HFunctor instance from any instance of HBifunctor, using -XDerivingVia.

For example, because we have instance HBifunctor Day, we can write:

deriving via (WrappedHBifunctor Day f) instance HFunctor (Day f)

to give us an automatic HFunctor instance and save us some work.

Constructors

WrapHBifunctor

Fields

unwrapHBifunctor :: t f g a

overHBifunctor :: HBifunctor t => (f <~> f') -> (g <~> g') -> t f g <~> t f' g'

Lift two isomorphisms on each side of a bifunctor to become an isomorphism between the two bifunctor applications.

Basically, if f and f' are isomorphic, and g and g' are isomorphic, then t f g is isomorphic to t f' g'.

Combinators

newtype JoinT t f a

Form an HFunctor by applying the same input twice to an HBifunctor.

Constructors

JoinT

Fields

runJoinT :: t f f a

newtype TannenT t p f g a

Form an HBifunctor by wrapping another HBifunctor in a HFunctor.

Constructors

TannenT

Fields

runTannenT :: t (p f g) a

newtype BiffT p s t f g a

Form an HBifunctor over two HFunctors.

Constructors

BiffT

Fields

runBiffT :: p (s f) (t g) a

newtype HClown t f g a

Form an HBifunctor over a HFunctor by ignoring the second argument.

Constructors

HClown

Fields

runHClown :: t f a

newtype HJoker t f g a

Form an HBifunctor over a HFunctor by ignoring the first argument.

Constructors

HJoker

Fields

runHJoker :: t g a