Data.HBifunctor.Associative

Description

This module provides tools for working with binary functor combinators that represent interpretable schemas.

These are types HBifunctor t that take two functors f and g and returns a new functor t f g, that "mixes together" f and g in some way.

The high-level usage of this is

biretract :: t f f ~> f

which lets you fully "mix" together the two input functors.

This class also associates each HBifunctor with its "semigroup functor combinator", so we can "squish together" repeated applications of t.

That is, an NonEmptyBy t f a is either:

f a

t f f a

t f (t f f) a

t f (t f (t f f)) a

.. etc.

which means we can have "list-like" schemas that represent multiple copies of f.

See Data.HBifunctor.Tensor for a version that also provides an analogy to inject, and a more flexible "squished" combinator ListBy that has an "empty" element.

Associative

class (HBifunctor t, Inject (NonEmptyBy t)) => Associative t where

An HBifunctor where it doesn't matter which binds first is Associative. Knowing this gives us a lot of power to rearrange the internals of our HFunctor at will.

For example, for the functor product:

data (f :*: g) a = f a :*: g a

We know that f :*: (g :*: h) is the same as (f :*: g) :*: h.

Formally, we can say that t enriches a the category of endofunctors with semigroup strcture: it turns our endofunctor category into a "semigroupoidal category".

Different instances of t each enrich the endofunctor category in different ways, giving a different semigroupoidal category.

Minimal complete definition

associating, appendNE, matchNE

Associated Types

type NonEmptyBy t :: (Type -> Type) -> Type -> Type

The "semigroup functor combinator" generated by t.

A value of type NonEmptyBy t f a is equivalent to one of:

f a

t f f a

t f (t f f) a

t f (t f (t f f)) a

t f (t f (t f (t f f))) a

.. etc

For example, for :*:, we have NonEmptyF. This is because:

x             ~ NonEmptyF (x :| [])      ~ inject x
x :*: y       ~ NonEmptyF (x :| [y])     ~ toNonEmptyBy (x :*: y)
x :*: y :*: z ~ NonEmptyF (x :| [y,z])
-- etc.

You can create an "singleton" one with inject, or else one from a single t f f with toNonEmptyBy.

Methods

associating :: (Functor f, Functor g, Functor h) => t f (t g h) <~> t (t f g) h

The isomorphism between t f (t g h) a and t (t f g) h a. To use this isomorphism, see assoc and disassoc.

appendNE :: t (NonEmptyBy t f) (NonEmptyBy t f) ~> NonEmptyBy t f

If a NonEmptyBy t f represents multiple applications of t f to itself, then we can also "append" two NonEmptyBy t fs applied to themselves into one giant NonEmptyBy t f containing all of the t fs.

matchNE :: Functor f => NonEmptyBy t f ~> (f :+: t f (NonEmptyBy t f))

If a NonEmptyBy t f represents multiple applications of t f to itself, then we can split it based on whether or not it is just a single f or at least one top-level application of t f.

Note that you can recursively "unroll" a NonEmptyBy completely into a Chain1 by using unrollNE.

consNE :: t f (NonEmptyBy t f) ~> NonEmptyBy t f

Prepend an application of t f to the front of a NonEmptyBy t f.

toNonEmptyBy :: t f f ~> NonEmptyBy t f

Embed a direct application of f to itself into a NonEmptyBy t f.

assoc :: (Associative t, Functor f, Functor g, Functor h) => t f (t g h) ~> t (t f g) h

Reassociate an application of t.

disassoc :: (Associative t, Functor f, Functor g, Functor h) => t (t f g) h ~> t f (t g h)

Reassociate an application of t.

SemigroupIn

class Associative t => SemigroupIn t f where

For different Associative t, we have functors f that we can "squash", using biretract:

t f f ~> f

This gives us the ability to squash applications of t.

Formally, if we have Associative t, we are enriching the category of endofunctors with semigroup structure, turning it into a semigroupoidal category. Different choices of t give different semigroupoidal categories.

A functor f is known as a "semigroup in the (semigroupoidal) category of endofunctors on t" if we can biretract:

t f f ~> f

This gives us a few interesting results in category theory, which you can stil reading about if you don't care:

All functors are semigroups in the semigroupoidal category on :+:

The class of functors that are semigroups in the semigroupoidal category on :*: is exactly the functors that are instances of Alt.

The class of functors that are semigroups in the semigroupoidal category on Day is exactly the functors that are instances of Apply.

The class of functors that are semigroups in the semigroupoidal category on Comp is exactly the functors that are instances of Bind.

Minimal complete definition

Nothing

Methods

biretract :: t f f ~> f

The HBifunctor analogy of retract. It retracts both fs into a single f, effectively fully mixing them together.

This function makes f a semigroup in the category of endofunctors with respect to tensor t.

default biretract :: Interpret (NonEmptyBy t) f => t f f ~> f

binterpret :: (g ~> f) -> (h ~> f) -> t g h ~> f

The HBifunctor analogy of interpret. It takes two interpreting functions, and mixes them together into a target functor h.

default binterpret :: Interpret (NonEmptyBy t) f => (g ~> f) -> (h ~> f) -> t g h ~> f

matchingNE :: (Associative t, Functor f) => NonEmptyBy t f <~> (f :+: t f (NonEmptyBy t f))

An NonEmptyBy t f represents the successive application of t to f, over and over again. So, that means that an NonEmptyBy t f must either be a single f, or an t f (NonEmptyBy t f).

matchingNE states that these two are isomorphic. Use matchNE and inject !*! consNE to convert between one and the other.

Utility

biget :: SemigroupIn t (Const b) => (forall x. f x -> b) -> (forall x. g x -> b) -> t f g a -> b

Useful wrapper over binterpret to allow you to directly extract a value b out of the t f a, if you can convert f x into b.

Note that depending on the constraints on f in SemigroupIn t f, you may have extra constraints on b.

If f is unconstrained, there are no constraints on b

If f must be Apply, b needs to be an instance of Semigroup

If f must be Applicative, b needs to be an instance of Monoid

For some constraints (like Monad), this will not be usable.

-- Return the length of either the list, or the Map, depending on which
--   one s in the +
biget length length
    :: ([] :+: Map Int) Char
    -> Int

-- Return the length of both the list and the map, added together
biget (Sum . length) (Sum . length)
    :: Day [] (Map Int) Char
    -> Sum Int

bicollect :: SemigroupIn t (Const [b]) => (forall x. f x -> b) -> (forall x. g x -> b) -> t f g a -> [b]

Useful wrapper over biget to allow you to collect a b from all instances of f and g inside a t f g a.

This will work if the constraint on f for SemigroupIn t f is Apply or Applicative, or if it is unconstrained.

(!*!) :: SemigroupIn t h => (f ~> h) -> (g ~> h) -> t f g ~> h infixr 5

Infix alias for binterpret

(!$!) :: SemigroupIn t (Const b) => (forall x. f x -> b) -> (forall x. g x -> b) -> t f g a -> b infixr 5

Infix alias for biget

-- Return the length of either the list, or the Map, depending on which
--   one s in the +
length !$! length
    :: ([] :+: Map Int) Char
    -> Int

-- Return the length of both the list and the map, added together
Sum . length !$! Sum . length
    :: Day [] (Map Int) Char
    -> Sum Int