Data.HBifunctor.Tensor

Description

This module provides tools for working with binary functor combinators.

Data.Functor.HFunctor deals with single functor combinators (transforming a single functor). This module provides tools for working with combinators that combine and mix two functors "together".

The binary analog of HFunctor is HBifunctor: we can map a structure-transforming function over both of the transformed functors.

The binary analog of Interpret is Tensor. If your combinator is an instance of Tensor, it means that you can "squish" both arguments together into an Inject. For example:

toListBy :: (f :*: f) a -> ListF f a
toListBy :: Comp f f a -> Free f a
toListBy :: Day f f a -> Ap f a

Tensor

class (Associative t, Inject (ListBy t)) => Tensor t i | t -> i where

An Associative HBifunctor can be a Tensor if there is some identity i where t i f and t f i are equivalent to just f.

That is, "enhancing" f with t i does nothing.

The methods in this class provide us useful ways of navigating a Tensor t with respect to this property.

The Tensor is essentially the HBifunctor equivalent of Inject, with intro1 and intro2 taking the place of inject.

Formally, we can say that t enriches a the category of endofunctors with monoid strcture: it turns our endofunctor category into a "monoidal category".

Different instances of t each enrich the endofunctor category in different ways, giving a different monoidal category.

Minimal complete definition

intro1, intro2, elim1, elim2, appendLB, splitNE, splittingLB

Associated Types

type ListBy t :: (Type -> Type) -> Type -> Type

The "monoidal functor combinator" induced by t.

A value of type ListBy t f a is equivalent to one of:

I a -- zero fs

f a -- one f

t f f a -- two fs

t f (t f f) a -- three fs

t f (t f (t f f)) a

t f (t f (t f (t f f))) a

.. etc

For example, for :*:, we have ListF. This is because:

Proxy         ~ ListF []         ~ nilLB @(:*:)
x             ~ ListF [x]        ~ inject x
x :*: y       ~ ListF [x,y]      ~ toListBy (x :*: y)
x :*: y :*: z ~ ListF [x,y,z]
-- etc.

You can create an "empty" one with nilLB, a "singleton" one with inject, or else one from a single t f f with toListBy.

Methods

intro1 :: f ~> t f i

Because t f (I t) is equivalent to f, we can always "insert" f into t f (I t).

This is analogous to inject from Inject, but for HBifunctors.

intro2 :: g ~> t i g

Because t (I t) g is equivalent to f, we can always "insert" g into t (I t) g.

This is analogous to inject from Inject, but for HBifunctors.

elim1 :: Functor f => t f i ~> f

Witnesses the property that i is the identity of t: t f i always leaves f unchanged, so we can always just drop the i.

elim2 :: Functor g => t i g ~> g

Witnesses the property that i is the identity of t: t i g always leaves g unchanged, so we can always just drop the i t.

appendLB :: t (ListBy t f) (ListBy t f) ~> ListBy t f

If a ListBy t f represents multiple applications of t f to itself, then we can also "append" two ListBy t fs applied to themselves into one giant ListBy t f containing all of the t fs.

splitNE :: NonEmptyBy t f ~> t f (ListBy t f)

Lets you convert an NonEmptyBy t f into a single application of f to ListBy t f.

Analogous to a function NonEmpty a -> (a, [a])

Note that this is not reversible in general unless we have Matchable t.

splittingLB :: ListBy t f <~> (i :+: t f (ListBy t f))

An ListBy t f is either empty, or a single application of t to f and ListBy t f (the "head" and "tail"). This witnesses that isomorphism.

To use this property, see nilLB, consLB, and unconsLB.

toListBy :: t f f ~> ListBy t f

Embed a direct application of f to itself into a ListBy t f.

fromNE :: NonEmptyBy t f ~> ListBy t f

NonEmptyBy t f is "one or more fs", and 'ListBy t f is "zero or more fs". This function lets us convert from one to the other.

This is analogous to a function NonEmpty a -> [a].

Note that because t is not inferrable from the input or output type, you should call this using -XTypeApplications:

fromNE @(:*:) :: NonEmptyF f a -> ListF f a
fromNE @Comp  :: Free1 f a -> Free f a

rightIdentity :: (Tensor t i, Functor f) => f <~> t f i

f is isomorphic to t f i: that is, i is the identity of t, and leaves f unchanged.

leftIdentity :: (Tensor t i, Functor g) => g <~> t i g

g is isomorphic to t i g: that is, i is the identity of t, and leaves g unchanged.

sumLeftIdentity :: f <~> (V1 :+: f)

leftIdentity (intro1 and elim1) for :+: actually does not require Functor. This is the more general version.

sumRightIdentity :: f <~> (f :+: V1)

rightIdentity (intro2 and elim2) for :+: actually does not require Functor. This is the more general version.

prodLeftIdentity :: f <~> (Proxy :*: f)

leftIdentity (intro1 and elim1) for :*: actually does not require Functor. This is the more general version.

prodRightIdentity :: g <~> (g :*: Proxy)

rightIdentity (intro2 and elim2) for :*: actually does not require Functor. This is the more general version.

MonoidIn

class (Tensor t i, SemigroupIn t f) => MonoidIn t i f where

This class effectively gives us a way to generate a value of f a based on an i a, for Tensor t i. Having this ability makes a lot of interesting functions possible when used with biretract from SemigroupIn that weren't possible without it: it gives us a "base case" for recursion in a lot of cases.

Essentially, we get an i ~> f, pureT, where we can introduce an f a as long as we have an i a.

Formally, if we have Tensor t i, we are enriching the category of endofunctors with monoid structure, turning it into a monoidal category. Different choices of t give different monoidal categories.

A functor f is known as a "monoid in the (monoidal) category of endofunctors on t" if we can biretract:

t f f ~> f

and also pureT:

i ~> f

This gives us a few interesting results in category theory, which you can stil reading about if you don't care:

All functors are monoids in the monoidal category on :+:

The class of functors that are monoids in the monoidal category on :*: is exactly the functors that are instances of Plus.

The class of functors that are monoids in the monoidal category on Day is exactly the functors that are instances of Applicative.

The class of functors that are monoids in the monoidal category on Comp is exactly the functors that are instances of Monad.

This is the meaning behind the common adage, "monads are just monoids in the category of endofunctors". It means that if you enrich the category of endofunctors to be monoidal with Comp, then the class of functors that are monoids in that monoidal category are exactly what monads are. However, the adage is a little misleading: there are many other ways to enrich the category of endofunctors to be monoidal, and Comp is just one of them. Similarly, the class of functors that are monoids in the category of endofunctors enriched by Day are Applicative.

Note that instances of this class are intended to be written with t and i to be fixed type constructors, and f to be allowed to vary freely:

instance Monad f => MonoidIn Comp Identity f

Any other sort of instance and it's easy to run into problems with type inference. If you want to write an instance that's "polymorphic" on tensor choice, use the WrapHBF and WrapF newtype wrappers over type variables, where the third argument also uses a type constructor:

instance MonoidIn (WrapHBF t) (WrapF i) (MyFunctor t i)

This will prevent problems with overloaded instances.

Minimal complete definition

Nothing

Methods

pureT :: i ~> f

If we have an i, we can generate an f based on how it interacts with t.

Specialized (and simplified), this type is:

pureT @Day   :: Applicative f => Identity a -> f a  -- pure
pureT @Comp  :: Monad f => Identity a -> f a        -- return
pureT @(:*:) :: Plus f => Proxy a -> f a            -- zero

Note that because t appears nowhere in the input or output types, you must always use this with explicit type application syntax (like pureT @Day)

Along with biretract, this function makes f a monoid in the category of endofunctors with respect to tensor t.

default pureT :: Interpret (ListBy t) f => i ~> f

nilLB :: forall t i f. Tensor t i => i ~> ListBy t f

Create the "empty ListBy".

If ListBy t f represents multiple applications of t f with itself, then nilLB gives us "zero applications of f".

Note that t cannot be inferred from the input or output type of nilLB, so this function must always be called with -XTypeApplications:

nilLB @Day :: Identity ~> Ap f
nilLB @Comp :: Identity ~> Free f
nilLB @(:*:) :: Proxy ~> ListF f

consLB :: Tensor t i => t f (ListBy t f) ~> ListBy t f

Lets us "cons" an application of f to the front of an ListBy t f.

unconsLB :: Tensor t i => ListBy t f ~> (i :+: t f (ListBy t f))

"Pattern match" on an ListBy t

An ListBy t f is either empty, or a single application of t to f and ListBy t f (the "head" and "tail")

This is analogous to the function uncons :: [a] -> Maybe (a, [a]).

retractLB :: forall t i f. MonoidIn t i f => ListBy t f ~> f

An implementation of retract that works for any instance of MonoidIn t i for ListBy t.

Can be useful as a default implementation if you already have MonoidIn implemented.

interpretLB :: forall t i g f. MonoidIn t i f => (g ~> f) -> ListBy t g ~> f

An implementation of interpret that works for any instance of MonoidIn t i for ListBy t.

Can be useful as a default implementation if you already have MonoidIn implemented.

Utility

inL :: forall t i f g. MonoidIn t i g => f ~> t f g

Convenient wrapper over intro1 that lets us introduce an arbitrary functor g to the right of an f.

You can think of this as an HBifunctor analogue of inject.

inR :: forall t i f g. MonoidIn t i f => g ~> t f g

Convenient wrapper over intro2 that lets us introduce an arbitrary functor f to the right of a g.

You can think of this as an HBifunctor analogue of inject.

outL :: (Tensor t Proxy, Functor f) => t f g ~> f

Convenient wrapper over elim1 that lets us drop one of the arguments of a Tensor for free, without requiring any extra constraints (like for binterpret).

See prodOutL for a version that does not require Functor f, specifically for :*:.

outR :: (Tensor t Proxy, Functor g) => t f g ~> g

Convenient wrapper over elim2 that lets us drop one of the arguments of a Tensor for free, without requiring any constraints (like for binterpret).

See prodOutR for a version that does not require Functor g, specifically for :*:.

newtype WrapF f a

A newtype wrapper meant to be used to define polymorphic MonoidIn instances. See documentation for MonoidIn for more information.

Please do not ever define an instance of MonoidIn "naked" on the third parameter:

instance MonidIn (WrapHBF t) (WrapF i) f

As that would globally ruin everything using WrapHBF.

Constructors

WrapF

Fields

unwrapF :: f a

prodOutL :: (f :*: g) ~> f

outL for :*: actually does not require Functor. This is the more general version.

prodOutR :: (f :*: g) ~> g

outR for :*: actually does not require Functor. This is the more general version.

Matchable

class Tensor t i => Matchable t i where

For some t, we have the ability to "statically analyze" the ListBy t and pattern match and manipulate the structure without ever interpreting or retracting. These are Matchable.

Methods

unsplitNE :: t f (ListBy t f) ~> NonEmptyBy t f

The inverse of splitNE. A consing of f to ListBy t f is non-empty, so it can be represented as an NonEmptyBy t f.

This is analogous to a function uncurry (:|) :: (a, [a]) -> NonEmpty a.

matchLB :: ListBy t f ~> (i :+: NonEmptyBy t f)

"Pattern match" on an ListBy t f: it is either empty, or it is non-empty (and so can be an NonEmptyBy t f).

This is analgous to a function nonEmpty :: [a] -> Maybe (NonEmpty a).

Note that because t cannot be inferred from the input or output type, you should use this with -XTypeApplications:

matchLB @Day :: Ap f a -> (Identity :+: Ap1 f) a

Note that you can recursively "unroll" a ListBy completely into a Chain by using unrollLB.

splittingNE :: Matchable t i => NonEmptyBy t f <~> t f (ListBy t f)

An NonEmptyBy t f is isomorphic to an f consed with an ListBy t f, like how a NonEmpty a is isomorphic to (a, [a]).

matchingLB :: forall t i f. Matchable t i => ListBy t f <~> (i :+: NonEmptyBy t f)

An ListBy t f is isomorphic to either the empty case (i) or the non-empty case (NonEmptyBy t f), like how [a] is isomorphic to Maybe (NonEmpty a).