Data.HFunctor

Description

This module provides abstractions for working with unary functor combinators.

Principally, it defines the HFunctor itself, as well as some classes that expose extra functionality that some HFunctors have (Inject and HBind).

See Data.HFunctor.Interpret for tools to use HFunctors as functor combinators that can represent interpretable schemas, and Data.HBifunctor for an abstraction over binary functor combinators.

Documentation

class HFunctor t where

An HFunctor can be thought of a unary "functor transformer" --- a basic functor combinator. It takes a functor as input and returns a functor as output.

It "enhances" a functor with extra structure (sort of like how a monad transformer enhances a Monad with extra structure).

As a uniform inteface, we can "swap the underlying functor" (also sometimes called "hoisting"). This is what hmap does: it lets us swap out the f in a t f for a t g.

For example, the free monad Free takes a Functor and returns a new Functor. In the process, it provides a monadic structure over f. hmap lets us turn a Free f into a Free g: a monad built over f can be turned into a monad built over g.

For the ability to move in and out of the enhanced functor, see Interpret.

Methods

hmap :: (f ~> g) -> t f ~> t g

If we can turn an f into a g, then we can turn a t f into a t g.

It must be the case that

hmap id == id

Essentially, t f adds some "extra structure" to f. hmap must swap out the functor, without affecting the added structure.

For example, ListF f a is essentially a list of f as. If we hmap to swap out the f as for g as, then we must ensure that the "added structure" (here, the number of items in the list) remains the same. So, hmap must preserve the number of items in the list.

The law hmap id == id is a way of formalizing this informal property.

overHFunctor :: HFunctor t => (f <~> g) -> t f <~> t g

Lift an isomorphism over an HFunctor.

Essentailly, if f and g are isomorphic, then so are t f and t g.

class HFunctor t => Inject t where

A typeclass for HFunctors where you can "inject" an f a into a t f a:

inject :: f a -> t f a

If you think of t f a as an "enhanced f", then inject allows you to use an f as its enhanced form.

inject itself is not too useful without Interpret to allow us to interpret or retrieve back the f.

Methods

inject :: f ~> t f

Lift an f into the enhanced t f structure. Analogous to lift from MonadTrans.

class Inject t => HBind t where

A typeclass for HFunctors that you can bind continuations to, without caring about what the contexts at play.

It is very similar to Interpret, except Interpret has the ability to constrain the contexts to some typeclass.

The main law is that binding inject should leave things unchanged:

hbind inject == id

But hbind should also be associatiatve, in a way that makes

hjoin . hjoin
   = hjoin . hmap hjoin

That is, squishing a t (t (t f)) a into a t f a can be done "inside" first, then "outside", or "outside" first, then "inside".

Minimal complete definition

hbind | hjoin

Methods

hbind :: (f ~> t g) -> t f ~> t g

Bind a continuation to a t f into some context g.

hjoin :: t (t f) ~> t f

Collapse a nested t (t f) into a single t f.

data HLift t f a

An "HFunctor combinator" that enhances an HFunctor with the ability to hold a single f a. This is the higher-order analogue of Lift.

This is mostly used as the semigroup fixed-point of ClownT.

Constructors

HPure (f a)

HOther (t f a)

data HFree t f a

An "HFunctor combinator" that turns an HFunctor into potentially infinite nestings of that HFunctor.

An HFree t f a is either f a, t f a, t (t f) a, t (t (t f)) a, etc.

This is the higher-oder analogue of Free.

This is mostly used as the semigroup fixed-point of JokerT, but is also the free HBind for any HFunctor t.

Constructors

HReturn (f a)

HJoin (t (HFree t f) a)