Data.HFunctor.Chain

Description

This module provides an Interpretable data type of "linked list of tensor applications".

The type Chain t, for any Monoidal t, is meant to be the same as ListBy t (the monoidal functor combinator for t), and represents "zero or more" applications of f to t.

The type Chain1 t, for any Semigroupoidal t, is meant to be the same as NonEmptyBy t (the semigroupoidal functor combinator for t) and represents "one or more" applications of f to t.

The advantage of using Chain and Chain1 over ListBy or NonEmptyBy is that they provide a universal interface for pattern matching and constructing such values, which may simplify working with new such functor combinators you might encounter.

Chain

data Chain t i f a

A useful construction that works like a "linked list" of t f applied to itself multiple times. That is, it contains t f f, t f (t f f), t f (t f (t f f)), etc, with f occuring zero or more times. It is meant to be the same as ListBy t.

If t is Monoidal, then it means we can "collapse" this linked list into some final type ListBy t (rerollLB), and also extract it back into a linked list (unrollLB).

So, a value of type Chain t i f a is one of either:

i a

f a

t f f a

t f (t f f) a

t f (t f (t f f)) a

.. etc.

Note that this is exactly what an ListBy t is supposed to be. Using Chain allows us to work with all ListBy ts in a uniform way, with normal pattern matching and normal constructors.

You can fully "collapse" a Chain t t f into an f with retract, if t is Monoidal; this could be considered a fundamental property of monoidal-ness.

This construction is inspired by http://oleg.fi/gists/posts/2018-02-21-single-free.html

Constructors

Done (i a)

More (t f (Chain t i f) a)

foldChain

Arguments

:: forall t i f g. HBifunctor t

=> (i ~> g)
Handle Done

-> (t f g ~> g)
Handle More

-> Chain t i f ~> g

Recursively fold down a Chain. Provide a function on how to handle the "single f case" (nilLB), and how to handle the "combined t f g case", and this will fold the entire Chain t i) f into a single g.

This is a catamorphism.

unfoldChain :: forall t f (g :: Type -> Type) i. HBifunctor t => (g ~> (i :+: t f g)) -> g ~> Chain t i f

Recursively build up a Chain. Provide a function that takes some starting seed g and returns either "done" (i) or "continue further" (t f g), and it will create a Chain t i f from a g.

This is an anamorphism.

unrollLB :: MonoidIn t i f => ListBy t f ~> Chain t i f

A type ListBy t is supposed to represent the successive application of ts to itself. unrollLB makes that successive application explicit, buy converting it to a literal Chain of applications of t to itself.

unrollLB = unfoldChain unconsLB

rerollLB :: forall t i f. MonoidIn t i f => Chain t i f ~> ListBy t f

A type ListBy t is supposed to represent the successive application of ts to itself. rerollNEB takes an explicit Chain of applications of t to itself and rolls it back up into an ListBy t.

rerollLB = foldChain nilLB consLB

Because t cannot be inferred from the input or output, you should call this with -XTypeApplications:

rerollLB @Comp
    :: Chain Comp Identity f a -> Free f a

unrollingLB :: MonoidIn t i f => ListBy t f <~> Chain t i f

A type ListBy t is supposed to represent the successive application of ts to itself. The type Chain t i f is an actual concrete ADT that contains successive applications of t to itself, and you can pattern match on each layer.

unrollingLB states that the two types are isormorphic. Use unrollLB and rerollLB to convert between the two.

appendChain :: forall t i f. MonoidIn t i f => t (Chain t i f) (Chain t i f) ~> Chain t i f

Chain is a monoid with respect to t: we can "combine" them in an associative way. The identity here is anything made with the Done constructor.

Since: 0.1.1.0

Chain1

data Chain1 t f a

A useful construction that works like a "non-empty linked list" of t f applied to itself multiple times. That is, it contains t f f, t f (t f f), t f (t f (t f f)), etc, with f occuring one or more times. It is meant to be the same as NonEmptyBy t.

A Chain1 t f a is explicitly one of:

f a

t f f a

t f (t f f) a

t f (t f (t f f)) a

.. etc

Note that this is exactly the description of NonEmptyBy t. And that's "the point": for all instances of Semigroupoidal, Chain1 t is isomorphic to NonEmptyBy t (witnessed by unrollingNEB). That's big picture of NonEmptyBy: it's supposed to be a type that consists of all possible self-applications of f to t.

Chain1 gives you a way to work with all NonEmptyBy t in a uniform way. Unlike for NonEmptyBy t f in general, you can always explicitly /pattern match/ on a Chain1 (with its two constructors) and do what you please with it. You can also construct Chain1 using normal constructors and functions.

You can convert in between NonEmptyBy t f and Chain1 t f with unrollNEB and rerollNEB. You can fully "collapse" a Chain1 t f into an f with retract, if t is Semigroupoidal; this could be considered a fundamental property of semigroupoidal-ness.

See Chain for a version that has an "empty" value.

This construction is inspired by iteratees and machines.

Constructors

Done1 (f a)

More1 (t f (Chain1 t f) a)

foldChain1

Arguments

:: forall t f g. HBifunctor t

=> (f ~> g)
handle Done1

-> (t f g ~> g)
handle More1

-> Chain1 t f ~> g

Recursively fold down a Chain1. Provide a function on how to handle the "single f case" (inject), and how to handle the "combined t f g case", and this will fold the entire Chain1 t f into a single g.

This is a catamorphism.

unfoldChain1 :: forall t f (g :: Type -> Type). HBifunctor t => (g ~> (f :+: t f g)) -> g ~> Chain1 t f

Recursively build up a Chain1. Provide a function that takes some starting seed g and returns either "done" (f) or "continue further" (t f g), and it will create a Chain1 t f from a g.

This is an anamorphism.

unrollingNEB :: forall t f. (SemigroupIn t f, Functor f) => NonEmptyBy t f <~> Chain1 t f

A type NonEmptyBy t is supposed to represent the successive application of ts to itself. The type Chain1 t f is an actual concrete ADT that contains successive applications of t to itself, and you can pattern match on each layer.

unrollingNEB states that the two types are isormorphic. Use unrollNEB and rerollNEB to convert between the two.

unrollNEB :: (SemigroupIn t f, Functor f) => NonEmptyBy t f ~> Chain1 t f

A type NonEmptyBy t is supposed to represent the successive application of ts to itself. unrollNEB makes that successive application explicit, buy converting it to a literal Chain1 of applications of t to itself.

unrollNEB = unfoldChain1 matchNEB

rerollNEB :: SemigroupIn t f => Chain1 t f ~> NonEmptyBy t f

A type NonEmptyBy t is supposed to represent the successive application of ts to itself. rerollNEB takes an explicit Chain1 of applications of t to itself and rolls it back up into an NonEmptyBy t.

rerollNEB = foldChain1 inject consNEB

appendChain1 :: forall t f. (SemigroupIn t f, Functor f) => t (Chain1 t f) (Chain1 t f) ~> Chain1 t f

Chain1 is a semigroup with respect to t: we can "combine" them in an associative way.

Since: 0.1.1.0

fromChain1 :: Tensor t i => Chain1 t f ~> Chain t i f

A Chain1 is "one or more linked fs", and a Chain is "zero or more linked fs". So, we can convert from a Chain1 to a Chain that always has at least one f.

The result of this function always is made with More at the top level.

Matchable

The following conversions between Chain and Chain1 are only possible if t is Matchable

splittingChain1 :: forall t i f. (MonoidIn t i f, Matchable t i, Functor f) => Chain1 t f <~> t f (Chain t i f)

A Chain1 t f is like a non-empty linked list of fs, and a Chain t i f is a possibly-empty linked list of fs. This witnesses the fact that the former is isomorphic to f consed to the latter.

splitChain1 :: forall t i f. MonoidIn t i f => Chain1 t f ~> t f (Chain t i f)

The "forward" function representing splittingChain1. Provided here as a separate function because it does not require Functor f.

matchingChain :: forall t i f. (MonoidIn t i f, Matchable t i, Functor f) => Chain t i f <~> (i :+: Chain1 t f)

A Chain t i f is a linked list of fs, and a Chain1 t f is a non-empty linked list of fs. This witnesses the fact that a Chain t i f is either empty (i) or non-empty (Chain1 t f).

unmatchChain :: forall t i f. MonoidIn t i f => (i :+: Chain1 t f) ~> Chain t i f

The "reverse" function representing matchingChain. Provided here as a separate function because it does not require Functor f.